+++ paper/paper.bib
@@ -42,6 +42,7 @@
   pages={1--22},
   year={1997},
   publisher={SIAM}
+  doi = {10.1137/s1064827594276424}
 }
 
 @article{gardner2022sundials,
@@ -97,6 +98,7 @@
   pages={436--451},
   year={1985},
   publisher={IEEE}
+  doi={10.1109/T-ED.1985.22232}
 }
 
 @article{hosea1996analysis,
@@ -108,6 +110,7 @@
   pages={21--37},
   year={1996},
   publisher={Elsevier}
+  doi={10.1016/0168-9274(95)00115-8}
 }
 
 @article{jorgensen2018family,
@@ -129,6 +132,7 @@ numpages = {18},
 keywords = {automatic differentiation, tasks, OpenMP, compiler, Julia, parallel, Enzyme, C++, RAJA, hybrid parallelization, MPI, distributed, LLVM},
 location = {Dallas, Texas},
 series = {SC '22}
+doi = {10.1109/SC41404.2022.00065}
 }
 
 @inproceedings{lattner2004llvm,
@@ -149,6 +153,7 @@ series = {SC '22}
   pages={770--775},
   year={2011},
   publisher={Elsevier}
+  doi={10.1016/j.camwa.2011.06.002}
 }
 
 @software{PyO3_Project_and_Contributors,
