+++ paper/paper.bib
@@ -41,7 +41,7 @@
   number={1},
   pages={1--22},
   year={1997},
-  publisher={SIAM}
+  publisher={SIAM},
   doi = {10.1137/s1064827594276424}
 }
 
@@ -97,7 +97,7 @@
   number={4},
   pages={436--451},
   year={1985},
-  publisher={IEEE}
+  publisher={IEEE},
   doi={10.1109/T-ED.1985.22232}
 }
 
@@ -109,7 +109,7 @@
   number={1-2},
   pages={21--37},
   year={1996},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/0168-9274(95)00115-8}
 }
 
@@ -131,7 +131,7 @@ articleno = {60},
 numpages = {18},
 keywords = {automatic differentiation, tasks, OpenMP, compiler, Julia, parallel, Enzyme, C++, RAJA, hybrid parallelization, MPI, distributed, LLVM},
 location = {Dallas, Texas},
-series = {SC '22}
+series = {SC '22},
 doi = {10.1109/SC41404.2022.00065}
 }
 
@@ -152,7 +152,7 @@ doi = {10.1109/SC41404.2022.00065}
   number={2},
   pages={770--775},
   year={2011},
-  publisher={Elsevier}
+  publisher={Elsevier},
   doi={10.1016/j.camwa.2011.06.002}
 }
 
