+++ paper/paper.bib
@@ -30,7 +30,8 @@
   number={3},
   pages={261--272},
   year={2020},
-  publisher={Nature Publishing Group US New York}
+  publisher={Nature Publishing Group US New York},
+  doi={10.1038/s41592-019-0686-2}
 }
 
 @article{shampine1997matlab,
@@ -76,6 +77,7 @@
     author={Patrick Kidger},
     year={2021},
     school={University of Oxford},
+    doi={h10.48550/arXiv.2202.02435}
 }
 
 @article{byrne1975polyalgorithm,
@@ -86,7 +88,8 @@
   number={1},
   pages={71--96},
   year={1975},
-  publisher={ACM New York, NY, USA}
+  publisher={ACM New York, NY, USA},
+  doi={10.1145/355626.355636}
 }
 
 @article{bank1985transient,
@@ -117,7 +120,8 @@
   title={A family of ESDIRK integration methods},
   author={J{\o}rgensen, John Bagterp and Kristensen, Morten Rode and Thomsen, Per Grove},
   journal={arXiv preprint arXiv:1803.01613},
-  year={2018}
+  year={2018},
+  doi= {10.48550/arXiv.1803.01613}
 }
 
 @inproceedings{moses22enzyme,
@@ -141,7 +145,8 @@ doi = {10.1109/SC41404.2022.00065}
   booktitle={International symposium on code generation and optimization, 2004. CGO 2004.},
   pages={75--86},
   year={2004},
-  organization={IEEE}
+  organization={IEEE},
+  doi={10.1109/CGO.2004.1281665}
 }
 
 @article{tsitouras2011runge,
