### 2025-11-06 9123580
infra: increment to macos-15
