### 2025-11-09 #9384 @editorialbot
[REVIEW]: `diffsol`: Rust crate for solving differential equations
**Submitting author:** @martinjrobins (<a href="http://orcid.org/0000-0002-1572-6782">Martin Robinson</a>)
**Repository:** https://github.com/martinjrobins/diffsol
**Branch with paper.md** (empty if default branch): joss
**Version:** v0.7.0
