### 2025-11-10 df60b2a
feat: support for f32 scalar types
