### 2025-11-10 dbee867
feat: support for f32 scalar types (#198)
