### 2025-11-10 ec9f197
infra: skip neural-ode weather prediction
