### 2025-11-10 265e470
feat: test on macos-15-intel (#177)
* infra: downgrade faer 0.22.6 -> 0.21.9, add x86 macos runner

* infra: test LLVM install via homebrew

* infra: increment to macos-15

* infra: skip neural-ode weather prediction

---------

Co-authored-by: Martin Robinson <martinjrobins@gmail.com>
