+++ examples/neural-ode-weather-prediction/Cargo.toml
@@ -4,12 +4,20 @@ version = "0.1.0"
 edition.workspace = true
 publish = false
 
+[features]
+default = []
+onnx = ["dep:ort", "dep:ort-sys"]
+
+[[bin]]
+name = "neural-ode-weather-prediction"
+required-features = ["onnx"]
+
 [dependencies]
 diffsol = { path = "../../diffsol" }
 nalgebra = { workspace = true }
 plotly = { workspace = true }
-ort = { workspace = true }
-ort-sys = { workspace = true }
+ort = { workspace = true, optional = true }
+ort-sys = { workspace = true, optional = true }
 ndarray = "0.16.1"
 csv = "1.3.1"
 rand = "0.9.0"
