### 2025-11-11 martinjrobins/diffsol#199 @jeertmans
[JOSS REVIEW] `diffsol`: Rust crate for solving differential equations
Hi @martinjrobins, thank you for submitting your work to JOSS :-)

xref: https://github.com/openjournals/joss-reviews/issues/9384

This piece of work looks very well-made, with a comprehensive documentation both for the API and the examples in the book.

Here are a few comments I had after reviewing the checklist.

> **Contribution and authorship:** Has the submitting author (@martinjrobins) made major contributions to the software? Does the full list of paper authors seem appropriate and complete?

Alex Allmont is cited as a co-author, but I cannot see any contribution from that person on this repository, nor to the paper. I see Alex Allmont is listed on your research group, but how has this person contributed to this tool / paper? I see this person is the maintainer of `pydiffsol`, but it is not directly the subject of the paper.

> **Installation instructions:** Is there a clearly-stated list of dependencies? Ideally these should be handled with an automated package management solution.

The [Installation and usage](https://github.com/martinjrobins/diffsol?tab=readme-ov-file#installation-and-usage) section points to the `crates.io` page, which itself is just a copy of the README, with no additional installation guidelines (except those provided by crates.io itself). I would recommend including a small sentence or two saying that you can simple `cargo add` your package. You can also include a dummy `cargo.toml` config file, that you can keep synchronized with tools like `cargo-release`. You also seem to provide optional feature flags, so it would be nice to document them (so people know what they do). [E.g., see how I did for my own tool](https://github.com/jeertmans/languagetool-rust?tab=readme-ov-file#feature-flags).

> **Example usage:** Do the authors include examples of how to use the software (ideally to solve real-world analysis problems).

I feel the paper and the README are missing a nice, simple example showcasing how you can solve, e.g., the Lorentz attractor, with your library (using the DSL?).

> **Automated tests:** Are there automated tests or manual steps described so that the functionality of the software can be verified?

There are tests (only inside the modules, no "integration tests" or tests within the book?), but I could not find documentation about how to run those tests.

> **Community guidelines:** Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support

I could not find any information about those three points. I think a short addendum to the README could solve these.

https://github.com/martinjrobins/diffsol/blob/4c6005bad05a32bbd9cc94752e7a61f02ba61ce9/paper/paper.md?plain=1#L55-L56

The `@byrne1975polyalgorithm` and `[@bank1985transient, @hosea1996analysis]` citations render a bit weirdly (in my opinion), as only the year is displayed between parentheses.

---

That's all for me, let me know if you have any question :-)
