### 2025-11-12 5576228
doc: add CONTRIBUTING and README example #199
