+++ README.md
@@ -1,5 +1,8 @@
 <div align="center">
-<p></p><img src="https://raw.githubusercontent.com/martinjrobins/diffsol/refs/heads/main/assets/diffsol_rectangle.svg" alt="Diffsol logo" width="300"/></p>
+<p></p><img src="https://raw.githubusercontent.com/martinjrobins/diffsol/refs/heads/main/assets/diffsol_rectangle.svg" alt="diffsol logo" width="300"/></p>
+<a href="https://martinjrobins.github.io/diffsol/">
+    <img src="https://img.shields.io/crates/v/diffsol.svg?label=mdbook&color=green&logo=mdbook" alt="mdbook badge">
+</a>
 <a href="https://docs.rs/diffsol">
     <img src="https://img.shields.io/crates/v/diffsol.svg?label=docs&color=blue&logo=rust" alt="docs.rs badge">
 </a>
@@ -10,7 +13,6 @@
     <img src="https://codecov.io/gh/martinjrobins/diffsol/branch/main/graph/badge.svg" alt="code coverage">
 </a>
 </div>
-
 ---
 
 Diffsol is a library for solving ordinary differential equations (ODEs) or semi-explicit differential algebraic equations (DAEs) in Rust. It can solve equations in the following form:
@@ -21,17 +23,68 @@ M \frac{dy}{dt} = f(t, y, p)
 
 where $M$ is a (possibly singular and optional) mass matrix, $y$ is the state vector, $t$ is the time and $p$ is a vector of parameters.
 
-The equations can be given by either rust code or the [DiffSL](https://martinjrobins.github.io/diffsl/) Domain Specific Language (DSL). The DSL uses automatic differentiation using [Enzyme](https://enzyme.mit.edu/) to calculate the necessary jacobians, and JIT compilation (using either [LLVM](https://llvm.org/) or [Cranelift](https://cranelift.dev/)) to generate efficient native code at runtime. The DSL is ideal for using Diffsol from a higher-level language like Python or R while still maintaining similar performance to pure rust.
+The equations can be given by either rust code or the [DiffSL](https://martinjrobins.github.io/diffsl/) Domain Specific Language (DSL). The DSL uses automatic differentiation using [Enzyme](https://enzyme.mit.edu/) to calculate the necessary jacobians, and JIT compilation (using either [LLVM](https://llvm.org/) or [Cranelift](https://cranelift.dev/)) to generate efficient native code at runtime. The DSL is ideal for using diffsol from a higher-level language like Python or R while still maintaining similar performance to pure rust.
 
-## Installation and Usage
+## Installation
 
-See installation instructions on the [crates.io page](https://crates.io/crates/diffsol).
+You can add diffsol using `cargo add diffsol` or directly in your `Cargo.toml`:
 
-The [Diffsol book](https://martinjrobins.github.io/diffsol/) describes how to use Diffsol using examples taken from several application areas (e.g. population dynamics, electrical circuits and pharmacological modelling), as well as more detailed information on the various APIs used to specify the ODE equations. For a more complete description of the API, please see the [docs.rs API documentation](https://docs.rs/diffsol).
+```toml
+[dependencies]
+diffsol = "0.7"
+```
 
-## Features
+Diffsol has the following features that can be enabled or disabled:
 
-The following solvers are available in Diffsol
+- `nalgebra`: Use nalgebra for linear algebra containers and solvers (enabled by default).
+- `faer`: Use faer for linear algebra containers and solvers (enabled by default).
+- `cuda`: Use in-built CUDA linear algebra containers and solvers (disabled by default, experimental).
[…]
