+++ CONTRIBUTING.md
@@ -1,0 +1,416 @@
+# Contributing to diffsol
+
+Thank you for your interest in contributing to diffsol! This guide will help you get started with contributing, and is written for someone new to contributions to a Rust crate. If you are already familiar with Rust development and contributing to open-source libraries on Github you can skip over most of the sections, however the "Setting Up Your Development Environment" section may still be useful as it includes information on optional dependencies required for certain features of diffsol.
+
+## Table of Contents
+
+- [Getting Started](#getting-started)
+  - [Forking the Repository](#forking-the-repository)
+  - [Setting Up Your Development Environment](#setting-up-your-development-environment)
+  - [Building the Project](#building-the-project)
+- [Making Changes](#making-changes)
+  - [Creating a Branch](#creating-a-branch)
+  - [Writing Code](#writing-code)
+  - [Code Style and Formatting](#code-style-and-formatting)
+- [Testing](#testing)
+  - [Running Tests](#running-tests)
+  - [Writing Tests](#writing-tests)
+  - [Code Coverage](#code-coverage)
+- [Documentation](#documentation)
+- [Submitting Your Contribution](#submitting-your-contribution)
+  - [Committing Your Changes](#committing-your-changes)
+  - [Opening a Pull Request](#opening-a-pull-request)
+- [Review Process](#review-process)
+- [Getting Help](#getting-help)
+
+## Getting Started
+
+### Diffsol repositories
+
+The diffsol ecosystem consists of several repositories. Before contributing, please ensure you route your issue or pull request to the appropriate repository:
+
+- [diffsol](https://github.com/martinjrobins/diffsol) - the core ODE solver library
+- [diffsl](https://github.com/martinjrobins/diffsl) - the DiffSL DSL compiler and JIT backends
+- [pydiffsol](https://github.com/alexallmont/pydiffsol) - Python bindings
+
+This guide focuses on contributing to the core diffsol library. Many of the principles apply to the other repositories as well, but please check their individual contributing guidelines for any differences.
+
+### Forking the Repository
+
+1. Navigate to the [diffsol repository](https://github.com/martinjrobins/diffsol) on GitHub.
+2. Click the **"Fork"** button in the upper right corner of the page.
+3. This creates a copy of the repository in your GitHub account.
+
+### Setting Up Your Development Environment
+
+#### Prerequisites
+
+- **Rust**: Install Rust using [rustup](https://rustup.rs/):
+
+  ```bash
+  curl --proto '=https' --tlsv1.2 -sSf https://sh.rustup.rs | sh
+  ```
+
+  After installation, ensure you have the latest stable version:
+
+  ```bash
+  rustup update stable
+  ```
+
+- **Git**: Ensure you have Git installed. You can download it from [git-scm.com](https://git-scm.com/).
+
+#### Clone Your Fork
+
+Clone your forked repository to your local machine:
+
+```bash
+git clone https://github.com/YOUR-USERNAME/diffsol.git
+cd diffsol
+```
+
+Replace `YOUR-USERNAME` with your GitHub username.
+
+#### Add Upstream Remote
+
+Add the original repository as an upstream remote to keep your fork synchronized:
+
+```bash
+git remote add upstream https://github.com/martinjrobins/diffsol.git
+```
+
+#### Install Optional Dependencies
+
+Diffsol has several optional features that may require additional dependencies:
+
+- **SuiteSparse KLU** (for `suitesparse` feature):
+  - On Ubuntu/Debian: `sudo apt-get install libsuitesparse-dev`
+  - On macOS: `brew install suite-sparse`
+
[…]
