+++ .github/workflows/rust.yml
@@ -101,6 +101,8 @@ jobs:
 
     steps:
     - uses: actions/checkout@v4
+    - name: Extract README code example
+      run: python3 ./.github/workflows/extract_readme_code.py
     - uses: actions/cache@v4
       with:
         path: |
