### 2025-11-12 3fb460d
doc: fix readme format
