### 2025-11-12 a87940e
doc: restructure readme
