+++ README.md
@@ -100,6 +100,17 @@ All solvers feature:
 - Forward sensitivity analysis, calculating the gradient of an output function or the solver states $y$ with respect to the parameters $p$.
 - Adjoint sensitivity analysis, calculating the gradient of cost function $G(p)$ with respect to the parameters $p$. The cost function can be the integral of a continuous output function $g(t, y, p)$ or a sum of a set of discrete functions $h_i(t_i, y_i, p)$ at time points $t_i$.
 
+## Contributing
+
+Contributions are very welcome, as are bug reports! Please see the [contributing guidelines](CONTRIBUTING.md) for more information, but in summary:
+
+- Please open an [issue](https://github.com/martinjrobins/diffsol/issues) or [discussion](https://github.com/martinjrobins/diffsol/discussions) to report any issues or problems using diffsol
+- Feel free to submit a pull request with your changes or improvements, but please open an issue first if the change is significant. The [contributing guidelines](CONTRIBUTING.md) describe how to set up a development environment, run tests, and format code.
+- There are a number of repositories in the diffsol ecosystem, please route your issue/request to the appropriate repository:
+  - [diffsol](https://github.com/martinjrobins/diffsol) - the core ODE solver library
+  - [diffsl](https://github.com/martinjrobins/diffsl) - the DiffSL DSL compiler and JIT backends
+  - [pydiffsol](https://github.com/alexallmont/pydiffsol) - Python bindings
+  
 ## Wanted - Developers for higher-level language wrappers
 
 Diffsol is designed to be easy to use from higher-level languages like Python or R. I'd prefer not to split my focus away from the core library, so I'm looking for developers who would like to lead the development of these wrappers. If you're interested, please get in touch.
@@ -112,17 +123,6 @@ Diffsol is designed to be easy to use from higher-level languages like Python or
 - [ ] Javascript in backend (e.g using [Neon](https://neon-rs.dev/))
 - [ ] Javascript in browser (e.g. using [wasm-pack](https://rustwasm.github.io/wasm-pack/))
 - [ ] Others, feel free to suggest your favourite language.
-
-## Contributing
-
-Contributions are very welcome, as are bug reports! Please see the [contributing guidelines](CONTRIBUTING.md) for more information, but in summary:
-
-- Please open an [issue](https://github.com/martinjrobins/diffsol/issues) or [discussion](https://github.com/martinjrobins/diffsol/discussions) to report any issues or problems using diffsol
-- Feel free to submit a pull request with your changes or improvements, but please open an issue first if the change is significant. The [contributing guidelines](CONTRIBUTING.md) describe how to set up a development environment, run tests, and format code.
-- There are a number of repositories in the diffsol ecosystem, please route your issue/request to the appropriate repository:
-  - [diffsol](https://github.com/martinjrobins/diffsol) - the core ODE solver library
-  - [diffsl](https://github.com/martinjrobins/diffsl) - the DiffSL DSL compiler and JIT backends
-  - [pydiffsol](https://github.com/alexallmont/pydiffsol) - Python bindings
 
 ## References
 
