### 2025-11-12 cc0bc63
doc: rearrange contributing bullets
