### 2025-11-12 martinjrobins/diffsol#199 @martinjrobins
Hi @jeertmans, I've made these changes in #200, see what you think (I don't think I can actually add you as a reviewer...?)

I'll do the changes to the paper in the `joss` branch....
