+++ examples/lorenz-attractor-diffsl-llvm/src/lorenz.rs
@@ -1,9 +1,8 @@
-use diffsol::{LlvmModule, OdeBuilder, OdeSolverMethod, NalgebraMat, NalgebraLU};
+use diffsol::{LlvmModule, NalgebraLU, NalgebraMat, OdeBuilder, OdeSolverMethod};
 
 pub fn lorenz() -> Result<(), Box<dyn std::error::Error>> {
-    let problem = OdeBuilder::<NalgebraMat<f64>>::new()
-        .build_from_diffsl::<LlvmModule>(
-            "
+    let problem = OdeBuilder::<NalgebraMat<f64>>::new().build_from_diffsl::<LlvmModule>(
+        "
             a { 14.0 } b { 10.0 } c { 8.0 / 3.0 }
             u_i {
                 x = 1.0,
@@ -15,9 +14,9 @@ pub fn lorenz() -> Result<(), Box<dyn std::error::Error>> {
                 x * (a - z) - y;
                 x * y - c * z;
             }
-        ")?;
+        ",
+    )?;
     let mut solver = problem.bdf::<NalgebraLU<f64>>()?;
     let (_ys, _ts) = solver.solve(0.0)?;
     Ok(())
 }
-
