### 2025-11-13 2e720f8
doc: more detailed README and adds CONTRIBUTING.md (#200)
* doc: add CONTRIBUTING and README example #199

* doc: fix readme format

* doc: restructure readme

* doc: rearrange contributing bullets

* chore: cargo fmt
