+++ examples/lorenz-attractor-diffsl-llvm/src/main.rs
@@ -1,0 +1,7 @@
+#[cfg(feature = "diffsl-llvm")]
+mod lorenz;
+
+fn main() {
+    #[cfg(feature = "diffsl-llvm")]
+    lorenz::lorenz().unwrap();
+}
