+++ examples/lorenz-attractor-diffsl-llvm/Cargo.toml
@@ -1,0 +1,15 @@
+[package]
+name = "lorenz-attractor-diffsl-llvm"
+version = "0.1.0"
+edition.workspace = true
+publish = false
+
+[features]
+diffsl-llvm15 = ["diffsol/diffsl-llvm15", "diffsl-llvm"]
+diffsl-llvm16 = ["diffsol/diffsl-llvm16", "diffsl-llvm"]
+diffsl-llvm17 = ["diffsol/diffsl-llvm17", "diffsl-llvm"]
+diffsl-llvm18 = ["diffsol/diffsl-llvm18", "diffsl-llvm"]
+diffsl-llvm = []
+
+[dependencies]
+diffsol = { path = "../../diffsol"}
