### 2025-11-13 67b6342
add pydiffsol to summary
