### 2025-11-13 martinjrobins/diffsol#199 @martinjrobins
I've made the edits to the paper in the `joss` branch: https://github.com/martinjrobins/diffsol/pull/196, which are:
- fit citation rendering
- add sentence on pydiffsol in summary
