### 2025-11-13 martinjrobins/diffsol#199 @jeertmans
Thanks for addressing my comments!
