### 2025-11-20 5e93b8e
choose correct spacing for heat equation example
