### 2025-11-20 447cd3d
remove `multilingual = false`
