+++ book/book.toml
@@ -1,7 +1,6 @@
 [book]
 authors = ["martinjrobins"]
 language = "en"
-multilingual = false
 src = "src"
 title = "Diffsol"
 
