### 2025-11-20 a9e729d
doc: choose correct spacing for heat equation example  (#202)
* choose correct spacing for heat equation example

* remove `multilingual = false`
