### 2025-12-03 8513dff
chore: update diffsl to 0.6.1
