### 2025-12-03 0c32cdb
chore: update diffsl to 0.6.1 (#205)
