+++ diffsol/Cargo.toml
@@ -25,13 +25,16 @@ diffsl-llvm15 = ["diffsl/llvm15-0", "diffsl", "diffsl-llvm"]
 diffsl-llvm16 = ["diffsl/llvm16-0", "diffsl", "diffsl-llvm"]
 diffsl-llvm17 = ["diffsl/llvm17-0", "diffsl", "diffsl-llvm"]
 diffsl-llvm18 = ["diffsl/llvm18-1", "diffsl", "diffsl-llvm"]
+diffsl-llvm19 = ["diffsl/llvm19-1", "diffsl", "diffsl-llvm"]
+diffsl-llvm20 = ["diffsl/llvm20-1", "diffsl", "diffsl-llvm"]
+diffsl-llvm21 = ["diffsl/llvm21-1", "diffsl", "diffsl-llvm"]
 
 [dependencies]
 nalgebra = { workspace = true }
 nalgebra-sparse = { version = "0.11.0", features = ["io"] }
 num-traits = "0.2.17"
 serde = { version = "1.0", features = ["derive"] }
-diffsl = { package = "diffsl", version = "0.6.0", optional = true, features = [
+diffsl = { package = "diffsl", version = "0.6.1", optional = true, features = [
     "rayon",
 ] }
 petgraph = "0.8.3"
