+++ README.md
@@ -31,7 +31,7 @@ You can add diffsol using `cargo add diffsol` or directly in your `Cargo.toml`:
 
 ```toml
 [dependencies]
-diffsol = "0.7"
+diffsol = "0.8"
 ```
 
 Diffsol has the following features that can be enabled or disabled:
