### 2025-12-06 f1f9ab4
deps(diffsl): version 0.6.2 (#206)
