### 2025-12-08 1ab6da8
version 0.8.3: diffsl version 0.6.3
