+++ diffsol/Cargo.toml
@@ -1,6 +1,6 @@
 [package]
 name = "diffsol"
-version = "0.8.3"
+version = "0.8.4"
 edition.workspace = true
 description = "A library for solving ordinary differential equations (ODEs) in Rust."
 license.workspace = true
@@ -34,7 +34,7 @@ nalgebra = { workspace = true }
 nalgebra-sparse = { version = "0.11.0", features = ["io"] }
 num-traits = "0.2.17"
 serde = { version = "1.0", features = ["derive"] }
-diffsl = { package = "diffsl", version = "0.6.3", optional = true, features = [
+diffsl = { package = "diffsl", version = "0.6.4", optional = true, features = [
     "rayon",
 ] }
 petgraph = "0.8.3"
