### 2025-12-08 72bf434
version 0.8.5: diffsl version 0.6.5 (#208)
