### 2025-12-09 caf1e18
feat: add optional backtracking line search to newton non-linear solver
