+++ diffsol/src/nonlinear_solver/line_search.rs
@@ -1,0 +1,175 @@
+use crate::{
+    error::{DiffsolError, NonLinearSolverError},
+    non_linear_solver_error,
+    nonlinear_solver::convergence::ConvergenceStatus,
+    Convergence, Scalar, Scale, Vector,
+};
+use num_traits::{FromPrimitive, One, Pow};
+
+/// Line search trait for nonlinear solvers
+/// The line search is used to find an optimal step size for the Newton iteration.
+/// The line search modifies the delta vector in place to scale it by the optimal step size
+/// The line search returns the norm of the modified delta vector
+/// The x vector is also modified in place to take the optimal step
+pub trait LineSearch<V: Vector>: Default {
+    /// Take the optimal step for the current iteration
+    fn take_optimal_step(
+        &mut self,
+        x: &mut V,
+        norm: &mut V::T,
+        delta: &mut V,
+        error_y: &V,
+        fun: &impl Fn(&V, &mut V),
+        linear_solver: &impl Fn(&mut V) -> Result<(), DiffsolError>,
+        convergence: &mut Convergence<V>,
+    ) -> Result<(), DiffsolError>;
+
+    /// Reset the line search state
+    fn reset(&mut self);
+}
+
+#[derive(Default)]
+pub struct NoLineSearch;
+
+impl<V: Vector> LineSearch<V> for NoLineSearch {
+    /// No line search implementation, simply takes the full step
+    fn take_optimal_step(
+        &mut self,
+        x: &mut V,
+        norm: &mut V::T,
+        delta: &mut V,
+        error_y: &V,
+        fun: &impl Fn(&V, &mut V),
+        linear_solver: &impl Fn(&mut V) -> Result<(), DiffsolError>,
+        convergence: &mut Convergence<V>,
+    ) -> Result<(), DiffsolError> {
+        //delta = f_at_n
+        fun(x, delta);
+
+        //delta = -delta_n
+        linear_solver(delta)?;
+
+        // xn = xn + delta_n
+        x.sub_assign(&*delta);
+
+        // norm
+        *norm = convergence.norm(delta, error_y);
+        Ok(())
+    }
+
+    fn reset(&mut self) {}
+}
+
+///
+/// Backtracking line search implementation
+///
+/// Parameters:
+/// - tau: step size reduction factor (0 < tau < 1), default 0.5
+/// - c: Armijo condition constant (0 < c < 1), default 1e-4
+/// - steptol: minimum step size, default eps^(2/3)
+/// - max_iter: maximum number of line search iterations, default 100
+/// - n_iters: number of line search iterations performed during last call
+///
+pub struct BacktrackingLineSearch<V: Vector> {
+    pub tau: V::T,
+    pub c: V::T,
+    pub steptol: V::T,
+    pub max_iter: usize,
+    pub n_iters: usize,
+    delta0: V,
+    x0: V,
+}
+
+impl<V: Vector> Default for BacktrackingLineSearch<V> {
+    fn default() -> Self {
+        Self {
+            tau: V::T::from_f64(0.5).unwrap(),
+            c: V::T::from_f64(1e-4).unwrap(),
+            steptol: V::T::EPSILON.pow(V::T::from_f64(2.0 / 3.0).unwrap()),
+            max_iter: 100,
+            n_iters: 0,
+            delta0: V::zeros(0, Default::default()),
+            x0: V::zeros(0, Default::default()),
+        }
+    }
+}
+
+/// Backtracking line search implementation, derived from backtracking line search algorithm
+/// in Sundials IDA solver (https://github.com/LLNL/sundials/blob/main/src/ida/ida_ic.c#L480)
[…]
