### 2025-12-09 be08c9b
feat: add optional backtracking line search to newton non-linear solver (#209)
* feat: add optional backtracking line search to newton non-linear solver

* fix test

* doc: fix url
