### 2025-12-09 7d8511a
feat: move solver config to problem
