+++ diffsol/src/lib.rs
@@ -214,7 +214,7 @@ pub use ode_solver::{
     method::AugmentedOdeSolverMethod,
     method::OdeSolverMethod,
     method::OdeSolverStopReason,
-    problem::OdeSolverProblem,
+    problem::{InitialConditionSolverOptions, OdeSolverOptions, OdeSolverProblem},
     sdirk::Sdirk,
     sdirk_state::RkState,
     sensitivities::SensitivitiesOdeSolverMethod,
