### 2025-12-09 45ba27c
reduce default max nonlinear solver iterations
