### 2025-12-09 924af0c
fix sdirk tests
