+++ diffsol/src/ode_solver/sdirk.rs
@@ -527,13 +527,13 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 4
-        number_of_steps: 29
+        number_of_steps: 30
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 116
+        number_of_nonlinear_solver_iterations: 120
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 118
+        number_of_calls: 122
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -547,14 +547,14 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 7
-        number_of_steps: 55
+        number_of_steps: 59
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 660
+        number_of_nonlinear_solver_iterations: 708
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 222
-        number_of_jac_muls: 446
+        number_of_calls: 238
+        number_of_jac_muls: 478
         number_of_matrix_evals: 2
         number_of_jac_adj_muls: 0
         "###);
@@ -566,14 +566,14 @@ mod test {
         let mut s = problem.esdirk34::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 3
-        number_of_steps: 13
+        number_of_linear_solver_setups: 4
+        number_of_steps: 14
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 109
+        number_of_nonlinear_solver_iterations: 106
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 111
+        number_of_calls: 108
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -587,13 +587,13 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 13
-        number_of_steps: 7
-        number_of_error_test_failures: 8
-        number_of_nonlinear_solver_iterations: 109
-        number_of_nonlinear_solver_fails: 3
+        number_of_steps: 8
+        number_of_error_test_failures: 7
+        number_of_nonlinear_solver_iterations: 102
+        number_of_nonlinear_solver_fails: 2
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 113
+        number_of_calls: 106
         number_of_jac_muls: 6
         number_of_matrix_evals: 2
         number_of_jac_adj_muls: 0
@@ -607,13 +607,13 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 5
[…]
