### 2025-12-09 1272b1a
feat: move solver config to problem (#210)
* feat: move solver config to problem

* reduce default max nonlinear solver iterations

* fix sdirk tests
