+++ diffsol/src/ode_solver/sdirk.rs
@@ -124,7 +124,13 @@ where
     ) -> Result<Self, DiffsolError> {
         Rk::<Eqn, M>::check_sdirk_rk(&tableau)?;
         let rk = Rk::new(problem, state, tableau)?;
-        let mut ret = Self::_new(rk, problem, linear_solver, true, SdirkConfig::default())?;
+        let mut ret = Self::_new(
+            rk,
+            problem,
+            linear_solver,
+            true,
+            SdirkConfig::new(&problem.ode_options),
+        )?;
         ret.nonlinear_solver.set_problem(ret.op.as_ref().unwrap());
         Ok(ret)
     }
@@ -178,7 +184,13 @@ where
     ) -> Result<Self, DiffsolError> {
         Rk::<Eqn, M>::check_sdirk_rk(&tableau)?;
         let rk = Rk::new_augmented(problem, state, tableau, &augmented_eqn)?;
-        let mut ret = Self::_new(rk, problem, linear_solver, true, SdirkConfig::default())?;
+        let mut ret = Self::_new(
+            rk,
+            problem,
+            linear_solver,
+            true,
+            SdirkConfig::new(&problem.ode_options),
+        )?;
 
         ret.s_op = if augmented_eqn.integrate_main_eqn() {
             ret.nonlinear_solver.set_problem(ret.op.as_ref().unwrap());
@@ -515,13 +527,13 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 4
-        number_of_steps: 29
+        number_of_steps: 30
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 116
+        number_of_nonlinear_solver_iterations: 120
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 118
+        number_of_calls: 122
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -535,14 +547,14 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 7
-        number_of_steps: 55
+        number_of_steps: 59
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 660
+        number_of_nonlinear_solver_iterations: 708
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 222
-        number_of_jac_muls: 446
+        number_of_calls: 238
+        number_of_jac_muls: 478
         number_of_matrix_evals: 2
         number_of_jac_adj_muls: 0
         "###);
@@ -554,14 +566,14 @@ mod test {
         let mut s = problem.esdirk34::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 3
-        number_of_steps: 13
+        number_of_linear_solver_setups: 4
+        number_of_steps: 14
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 109
+        number_of_nonlinear_solver_iterations: 106
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
[…]

[…]
