### 2025-12-10 ef8b264
fix: convergence for line search
