### 2025-12-10 5845454
fix sens ic consistent
