### 2025-12-10 7c940b7
tests: consistent initialisation
