### 2025-12-10 144551c
fix: line search and sens_equations sparsity (#211)
* fix: convergence for line search

* remove prints

* fix sens consistent ic as well

* fix sens ic consistent

* tests: consistent initialisation
