+++ diffsol/src/ode_solver/bdf.rs
@@ -1442,11 +1442,11 @@ mod test {
         number_of_linear_solver_setups: 11
         number_of_steps: 47
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 82
+        number_of_nonlinear_solver_iterations: 68
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 84
+        number_of_calls: 70
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -1488,11 +1488,11 @@ mod test {
         number_of_linear_solver_setups: 11
         number_of_steps: 47
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 82
+        number_of_nonlinear_solver_iterations: 68
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 84
+        number_of_calls: 70
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -1508,15 +1508,22 @@ mod test {
         number_of_linear_solver_setups: 13
         number_of_steps: 48
         number_of_error_test_failures: 1
-        number_of_nonlinear_solver_iterations: 272
+        number_of_nonlinear_solver_iterations: 236
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.statistics(), @r###"
-        number_of_calls: 89
-        number_of_jac_muls: 189
+        number_of_calls: 72
+        number_of_jac_muls: 170
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
         "###);
+    }
+
+    #[test]
+    fn bdf_test_faer_sparse_exponential_decay_sens() {
+        let (problem, soln) = exponential_decay_problem_sens::<FaerSparseMat<f64>>(false);
+        let mut s = problem.bdf_sens::<FaerSparseLU<f64>>().unwrap();
+        test_ode_solver(&mut s, soln, None, false, true);
     }
 
     #[cfg(feature = "diffsl-llvm")]
@@ -1530,7 +1537,7 @@ mod test {
         number_of_linear_solver_setups: 13
         number_of_steps: 48
         number_of_error_test_failures: 1
-        number_of_nonlinear_solver_iterations: 272
+        number_of_nonlinear_solver_iterations: 236
         number_of_nonlinear_solver_fails: 0
         "###);
     }
@@ -1559,10 +1566,10 @@ mod test {
             .unwrap();
         test_adjoint(adjoint_solver, dgdu);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 202
+        number_of_calls: 181
         number_of_jac_muls: 6
         number_of_matrix_evals: 3
-        number_of_jac_adj_muls: 425
+        number_of_jac_adj_muls: 376
         "###);
     }
 
@@ -1581,10 +1588,10 @@ mod test {
             .unwrap();
         test_adjoint_sum_squares(adjoint_solver, dgdp, soln, data, times.as_slice());
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 493
+        number_of_calls: 484
         number_of_jac_muls: 6
         number_of_matrix_evals: 3
-        number_of_jac_adj_muls: 1554
+        number_of_jac_adj_muls: 1481
         "###);
     }
 
@@ -1633,10 +1640,10 @@ mod test {
[…]
