### 2025-12-12 ee6dd9c
perf: add pybamm dfn benchmark
