### 2025-12-14 #9384 @DiogoRibeiro7
@nkrusch all completed and from my point of view that paper is ready to publish
