### 2025-12-15 95d9e05
feat: expose DiffSlScalar
- Required in python bindings when using build_from_diffsl with f32.
