### 2025-12-18 martinjrobins/diffsol#215 @Armavica
Broken link in the documentation
Hi, as part of the review that I am doing for JOSS (https://github.com/openjournals/joss-reviews/issues/9384), I am going through the diffsol book.

I noticed a broken link on the page https://martinjrobins.github.io/diffsol/primer/population_dynamics.html (on `Rust closures`)
