### 2025-12-18 martinjrobins/diffsol#216 @Armavica
Compiling the prey-predator fitting example
Hi, as part of the review that I am doing for JOSS (https://github.com/openjournals/joss-reviews/issues/9384) I am going through the diffsol book.

I am probably missing something obvious, but for some reason I can't get to compile this example https://martinjrobins.github.io/diffsol/primer/population_dynamics_fitting.html

I created a new rust project, copied all the code of the page inside `main.rs` and created the following `Cargo.toml`:

``` toml
[package]
name = "testdiffsol"
version = "0.1.0"
edition = "2024"

[dependencies]
diffsol = { version = "0.8.4", features = ["diffsl-llvm19"] }
nalgebra = "0.34.1"
plotly = "0.13.5"
argmin = "0.11.0"
argmin-math = "0.5.1"
argmin-observer-slog = "0.2.0"
```

What am I doing wrong?
