### 2025-12-18 martinjrobins/diffsol#216 @martinjrobins
what is the error that you are getting and your system details (os, llvm version / location, environment variables)? Since this example uses the llvm backend, compilation is more tricky than normal, you need to make sure you set the right env variables to point to your local install, see the README for more details

You can see a complete project with this example here: https://github.com/martinjrobins/diffsol/tree/main/examples/predator-prey-fitting-forward
