### 2025-12-18 fedb099
include solver creation in bench, update diffsl
