+++ diffsol/Cargo.toml
@@ -34,7 +34,7 @@ nalgebra = { workspace = true }
 nalgebra-sparse = { version = "0.11.0", features = ["io"] }
 num-traits = "0.2.17"
 serde = { version = "1.0", features = ["derive"] }
-diffsl = { package = "diffsl", version = "0.6.5", optional = true, features = [
+diffsl = { package = "diffsl", version = "0.7.1", optional = true, features = [
     "rayon",
 ] }
 petgraph = "0.8.3"
