### 2025-12-27 e6dffdd
diffsl calculates non-zeros now
