### 2025-12-28 e873a2f
feat: pybamm dfn bench, use diffsl symbolic sparsity (#218)
* perf: add pybamm dfn benchmark

* include solver creation in bench, update diffsl

* diffsl calculates non-zeros now

* diffsl 0.8.2
