+++ diffsol/src/ode_equations/diffsl.rs
@@ -1,17 +1,20 @@
 use core::panic;
 use num_traits::{One, Zero};
-use std::cell::RefCell;
 use std::ops::MulAssign;
+use std::{cell::RefCell, mem};
 
 use diffsl::{
-    execution::module::{CodegenModule, CodegenModuleCompile, CodegenModuleJit},
-    execution::scalar::Scalar as DiffSlScalar,
+    discretise::DiscreteModel,
+    execution::{
+        module::{CodegenModule, CodegenModuleCompile, CodegenModuleJit},
+        scalar::Scalar as DiffSlScalar,
+    },
+    parser::parse_ds_string,
     Compiler,
 };
 
 use crate::{
-    error::DiffsolError, find_jacobian_non_zeros, find_matrix_non_zeros, find_sens_non_zeros,
-    jacobian::JacobianColoring, matrix::sparsity::MatrixSparsity,
+    error::DiffsolError, jacobian::JacobianColoring, matrix::sparsity::MatrixSparsity,
     op::nonlinear_op::NonLinearOpJacobian, ConstantOp, ConstantOpSens, ConstantOpSensAdjoint,
     LinearOp, LinearOpTranspose, Matrix, MatrixHost, NonLinearOp, NonLinearOpAdjoint,
     NonLinearOpSens, NonLinearOpSensAdjoint, OdeEquations, OdeEquationsRef, Op, Scale, Vector,
@@ -38,6 +41,9 @@ pub struct DiffSlContext<M: Matrix<T: DiffSlScalar>, CG: CodegenModule> {
     has_out: bool,
     nout: usize,
     ctx: M::C,
+    rhs_state_deps: Vec<(usize, usize)>,
+    rhs_input_deps: Vec<(usize, usize)>,
+    mass_state_deps: Vec<(usize, usize)>,
 }
 
 impl<M: Matrix<T: DiffSlScalar>, CG: CodegenModuleCompile + CodegenModuleJit> DiffSlContext<M, CG> {
@@ -55,8 +61,15 @@ impl<M: Matrix<T: DiffSlScalar>, CG: CodegenModuleCompile + CodegenModuleJit> Di
             1 => diffsl::execution::compiler::CompilerMode::SingleThreaded,
             _ => diffsl::execution::compiler::CompilerMode::MultiThreaded(Some(nthreads)),
         };
-        let compiler = Compiler::from_discrete_str(text, mode)
-            .map_err(|e| DiffsolError::Other(e.to_string()))?;
+        let model =
+            parse_ds_string(text).map_err(|e| DiffsolError::DiffslParserError(e.to_string()))?;
+        let mut model = DiscreteModel::build("diffsol", &model)
+            .map_err(|e| DiffsolError::DiffslCompilerError(e.as_error_message(text)))?;
+        let compiler = Compiler::from_discrete_model(&model, mode)
+            .map_err(|e| DiffsolError::DiffslCompilerError(e.to_string()))?;
+        let rhs_state_deps = model.take_rhs_state_deps();
+        let rhs_input_deps = model.take_rhs_input_deps();
+        let mass_state_deps = model.take_mass_state_deps();
         let (nstates, _nparams, _nout, _ndata, _nroots, _has_mass) = compiler.get_dims();
 
         let compiler = if nthreads == 0 {
@@ -100,6 +113,9 @@ impl<M: Matrix<T: DiffSlScalar>, CG: CodegenModuleCompile + CodegenModuleJit> Di
             has_root,
             has_out,
             ctx,
+            rhs_state_deps,
+            rhs_input_deps,
+            mass_state_deps,
         })
     }
 }
[…]
