### 2025-12-29 4820692
feat: expose more solver options
