### 2025-12-29 5c557a8
feat: expose more solver options (#219)
