### 2025-12-29 7130265
perf: remove convergence prediction
