### 2025-12-29 71d20d8
perf: use average rate with convergence
