+++ diffsol/src/nonlinear_solver/convergence.rs
@@ -69,7 +69,7 @@ impl<'a, V: Vector> Convergence<'a, V> {
             return ConvergenceStatus::Converged;
         }
         if let Some(old_norm) = self.old_norm {
-            let rate = norm / old_norm;
+            let rate = (norm / old_norm).pow(V::T::one() / (V::T::from_usize(self.niter - 1).unwrap()));
             self.rate = Some(rate);
 
             // check if iteration is diverging
@@ -97,13 +97,13 @@ impl<'a, V: Vector> Convergence<'a, V> {
             if V::T::from_f64(100.0).unwrap() * norm < self.tol {
                 return ConvergenceStatus::Converged;
             }
+            self.old_norm = Some(norm);
         };
         ConvergenceStatus::Continue
     }
 
     pub fn check_new_iteration(&mut self, norm: V::T) -> ConvergenceStatus {
         let status = self.check_norm(norm);
-        self.old_norm = Some(norm);
         status
     }
 }
