### 2025-12-29 da8c5e7
perf: remove convergence prediction
