### 2026-01-02 4863fb3
feat: expose DiffSlScalar (#213)
- Required in python bindings when using build_from_diffsl with f32.
