### 2026-01-03 05c8155
reinstate the convergence prediction
