### 2026-01-05 069478e
sdirk works again, foodweb not starting for some reason
