+++ diffsol/src/ode_solver/sdirk.rs
@@ -11,10 +11,12 @@ use crate::Tableau;
 use crate::{
     nonlinear_solver::NonLinearSolver, op::sdirk::SdirkCallable, AugmentedOdeEquations,
     AugmentedOdeEquationsImplicit, Convergence, DefaultDenseMatrix, DenseMatrix, JacobianUpdate,
-    NoLineSearch, OdeEquationsImplicit, OdeSolverMethod, OdeSolverProblem, OdeSolverState, Op,
-    StateRef, StateRefMut,
+    Matrix, NoLineSearch, OdeEquationsImplicit, OdeSolverMethod, OdeSolverProblem, OdeSolverState,
+    Op, StateRef, StateRefMut,
 };
-use num_traits::{FromPrimitive, One};
+use log::debug;
+use log::trace;
+use num_traits::{FromPrimitive, One, ToPrimitive, Zero};
 
 use super::bdf::BdfStatistics;
 use super::config::SdirkConfig;
@@ -211,19 +213,15 @@ where
 
     fn jacobian_updates(&mut self, h: Eqn::T, state: SolverState) {
         if self.jacobian_update.check_rhs_jacobian_update(h, &state) {
-            println!("Updating RHS Jacobian at time {}, step size {}", self.rk.state().t, h);
             if let Some(op) = self.op.as_mut() {
                 op.set_jacobian_is_stale();
-                self.nonlinear_solver.reset_jacobian(
-                    op,
-                    &self.rk.old_state().dy,
-                    self.rk.state().t,
-                );
+                self.nonlinear_solver
+                    .reset_jacobian(op, &self.rk.state().y, self.rk.state().t);
             } else if let Some(s_op) = self.s_op.as_mut() {
                 s_op.set_jacobian_is_stale();
                 self.nonlinear_solver.reset_jacobian(
                     s_op,
-                    &self.rk.old_state().ds[0],
+                    &self.rk.state().s[0],
                     self.rk.state().t,
                 );
             }
@@ -231,18 +229,14 @@ where
             self.jacobian_update.update_jacobian(h);
             self.convergence.reset_eta();
         } else if self.jacobian_update.check_jacobian_update(h, &state) {
-            println!("Updating Jacobian at time {}, step size {}", self.rk.state().t, h);
             // shouldn't matter what we put in for x cause rhs_jacobian is already updated
             if let Some(op) = self.op.as_ref() {
-                self.nonlinear_solver.reset_jacobian(
-                    op,
-                    &self.rk.old_state().dy,
-                    self.rk.state().t,
-                );
+                self.nonlinear_solver
+                    .reset_jacobian(op, &self.rk.state().y, self.rk.state().t);
             } else if let Some(s_op) = self.s_op.as_ref() {
                 self.nonlinear_solver.reset_jacobian(
                     s_op,
-                    &self.rk.old_state().ds[0],
+                    &self.rk.state().s[0],
                     self.rk.state().t,
                 );
             }
@@ -334,6 +328,11 @@ where
     }
 
     fn step(&mut self) -> Result<OdeSolverStopReason<Eqn::T>, DiffsolError> {
+        debug!(
+            "Taking SDIRK step at time {}, step size {}",
+            self.rk.state().t,
+            self.rk.state().h
+        );
         let mut h = self.rk.start_step()?;
 
         // setup the operators for the step
@@ -348,6 +347,7 @@ where
             self.rk
                 .start_step_attempt(h, self.s_op.as_mut().map(|s_op| s_op.eqn_mut()));
             for i in start..self.rk.tableau().s() {
+                trace!("SDIRK Stage {}:", i);
                 if self
                     .rk
                     .do_stage_sdirk(
@@ -362,25 +362,52 @@ where
                 {
[…]
