+++ diffsol/src/error.rs
@@ -96,10 +96,12 @@ pub enum OdeSolverError {
     StateNotSet,
     #[error("Sensitivity solve failed")]
     SensitivitySolveFailed,
-    #[error("Exceeded maximum number of nonlinear solver failures at time = {time}")]
-    TooManyNonlinearSolverFailures { time: f64 },
-    #[error("Exceeded maximum number of error test failures at time = {time}")]
-    TooManyErrorTestFailures { time: f64 },
+    #[error(
+        "Exceeded maximum number of nonlinear solver failures ({num_failures}) at time = {time}"
+    )]
+    TooManyNonlinearSolverFailures { time: f64, num_failures: usize },
+    #[error("Exceeded maximum number of error test failures ({num_failures}) at time = {time}")]
+    TooManyErrorTestFailures { time: f64, num_failures: usize },
     #[error("Step size is too small at time = {time}")]
     StepSizeTooSmall { time: f64 },
     #[error("Sensitivity requested but equations do not support it")]
