+++ diffsol/src/ode_solver/bdf.rs
@@ -1,7 +1,7 @@
-use log::debug;
+use log::{debug, info, trace};
 use nalgebra::ComplexField;
-use std::cell::Ref;
 use std::ops::AddAssign;
+use std::{cell::Ref, fmt::Display};
 
 use crate::{
     error::{DiffsolError, OdeSolverError},
@@ -32,6 +32,12 @@ pub struct BdfStatistics {
     pub number_of_nonlinear_solver_fails: usize,
 }
 
+impl Display for BdfStatistics {
+    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
+        write!(f, "{}", serde_json::to_string_pretty(self).unwrap())
+    }
+}
+
 impl<'a, M, Eqn, Nls, AugEqn> AugmentedOdeSolverMethod<'a, Eqn, AugEqn>
     for Bdf<'a, Eqn, Nls, M, AugEqn>
 where
@@ -43,11 +49,16 @@ where
     Nls: NonLinearSolver<Eqn::M>,
     Eqn::V: DefaultDenseMatrix<T = Eqn::T>,
 {
-    fn into_state_and_eqn(self) -> (Self::State, Option<AugEqn>) {
-        (self.state, self.s_op.map(|op| op.eqn))
+    fn into_state_and_eqn(mut self) -> (Self::State, Option<AugEqn>) {
+        let callable = std::mem::take(&mut self.s_op);
+        let state = self.into_state();
+        (state, callable.map(|op| op.eqn))
     }
     fn augmented_eqn(&self) -> Option<&AugEqn> {
         self.s_op.as_ref().map(|op| op.eqn())
+    }
+    fn augmented_eqn_mut(&mut self) -> Option<&mut AugEqn> {
+        self.s_op.as_mut().map(|op| &mut op.eqn)
     }
 }
 
@@ -159,7 +170,7 @@ where
             alpha: self.alpha.clone(),
             gamma: self.gamma.clone(),
             error_const2: self.error_const2.clone(),
-            statistics: self.statistics.clone(),
+            statistics: BdfStatistics::default(),
             state: self.state.clone(),
             tstop: self.tstop,
             root_finder: self.root_finder.clone(),
@@ -167,6 +178,19 @@ where
             jacobian_update: self.jacobian_update.clone(),
             config: self.config.clone(),
         }
+    }
+}
+
+impl<'a, M, Eqn, Nls, AugmentedEqn> Drop for Bdf<'a, Eqn, Nls, M, AugmentedEqn>
+where
+    Eqn: OdeEquationsImplicit,
+    Nls: NonLinearSolver<Eqn::M>,
+    M: DenseMatrix<T = Eqn::T, V = Eqn::V, C = Eqn::C>,
+    AugmentedEqn: AugmentedOdeEquationsImplicit<Eqn>,
+    Eqn::V: DefaultDenseMatrix,
+{
+    fn drop(&mut self) {
+        info!("BDF Solver Statistics: {}", self.statistics);
     }
 }
 
@@ -481,6 +505,7 @@ where
         }
 
         self.state.h = new_h;
+        self.convergence.reset_eta_timestep_change();
 
         // if step size too small, then fail
         if self.state.h.abs() < self.config.minimum_timestep {
@@ -489,7 +514,6 @@ where
             }));
         }
 
-        self.convergence.reset_eta_timestep_change();
         Ok(new_h)
     }
 
@@ -658,15 +682,32 @@ where
         if (state.h > M::T::zero() && state.t + state.h > tstop + troundoff)
             || (state.h < M::T::zero() && state.t + state.h < tstop - troundoff)
         {
+            debug!(
+                "Adjusting step size to hit tstop at t = {}",
+                tstop.to_f64().unwrap()
+            );
[…]
