+++ diffsol/src/ode_solver/problem.rs
@@ -46,7 +46,7 @@ pub struct OdeSolverOptions<T: Scalar> {
 impl<T: Scalar> Default for OdeSolverOptions<T> {
     fn default() -> Self {
         Self {
-            max_nonlinear_solver_iterations: 4,
+            max_nonlinear_solver_iterations: 5,
             max_error_test_failures: 40,
             max_nonlinear_solver_failures: 50,
             min_timestep: T::from_f64(1e-13).unwrap(),
