### 2026-01-05 38565d3
make ts growth/shrink more consistent
