+++ diffsol/src/ode_solver/sdirk.rs
@@ -585,14 +585,14 @@ mod test {
         let mut s = problem.tr_bdf2::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 7
-        number_of_steps: 42
+        number_of_linear_solver_setups: 8
+        number_of_steps: 32
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 100
+        number_of_nonlinear_solver_iterations: 111
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 102
+        number_of_calls: 113
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -606,14 +606,14 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 9
-        number_of_steps: 75
+        number_of_steps: 55
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 511
+        number_of_nonlinear_solver_iterations: 516
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 166
-        number_of_jac_muls: 353
+        number_of_calls: 202
+        number_of_jac_muls: 322
         number_of_matrix_evals: 2
         number_of_jac_adj_muls: 0
         "###);
@@ -625,14 +625,14 @@ mod test {
         let mut s = problem.esdirk34::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 6
+        number_of_linear_solver_setups: 7
         number_of_steps: 15
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 55
+        number_of_nonlinear_solver_iterations: 76
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 57
+        number_of_calls: 78
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -645,14 +645,14 @@ mod test {
         let mut s = problem.esdirk34::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 5
+        number_of_linear_solver_setups: 6
         number_of_steps: 6
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 23
+        number_of_nonlinear_solver_iterations: 24
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 27
+        number_of_calls: 28
         number_of_jac_muls: 6
         number_of_matrix_evals: 2
         number_of_jac_adj_muls: 0
@@ -666,14 +666,14 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
[…]
