### 2026-01-05 8a6ef8c
remove rhs call in diffsl jacobian
