### 2026-01-05 bdac376
try tightening nls tolerance
