+++ diffsol/src/ode_solver/bdf.rs
@@ -1600,12 +1600,12 @@ mod test {
         number_of_linear_solver_setups: 14
         number_of_steps: 52
         number_of_error_test_failures: 1
-        number_of_nonlinear_solver_iterations: 163
+        number_of_nonlinear_solver_iterations: 165
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.statistics(), @r###"
         number_of_calls: 56
-        number_of_jac_muls: 113
+        number_of_jac_muls: 115
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
         "###);
@@ -1629,7 +1629,7 @@ mod test {
         number_of_linear_solver_setups: 14
         number_of_steps: 52
         number_of_error_test_failures: 1
-        number_of_nonlinear_solver_iterations: 163
+        number_of_nonlinear_solver_iterations: 165
         number_of_nonlinear_solver_fails: 0
         "###);
     }
@@ -1658,10 +1658,10 @@ mod test {
             .unwrap();
         test_adjoint(adjoint_solver, dgdu);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 160
+        number_of_calls: 161
         number_of_jac_muls: 6
         number_of_matrix_evals: 3
-        number_of_jac_adj_muls: 219
+        number_of_jac_adj_muls: 222
         "###);
     }
 
@@ -1683,7 +1683,7 @@ mod test {
         number_of_calls: 496
         number_of_jac_muls: 6
         number_of_matrix_evals: 3
-        number_of_jac_adj_muls: 1048
+        number_of_jac_adj_muls: 1065
         "###);
     }
 
@@ -1735,7 +1735,7 @@ mod test {
         number_of_calls: 166
         number_of_jac_muls: 18
         number_of_matrix_evals: 6
-        number_of_jac_adj_muls: 111
+        number_of_jac_adj_muls: 114
         "###);
     }
 
@@ -1757,7 +1757,7 @@ mod test {
         number_of_calls: 225
         number_of_jac_muls: 18
         number_of_matrix_evals: 6
-        number_of_jac_adj_muls: 411
+        number_of_jac_adj_muls: 421
         "###);
     }
 
@@ -1813,12 +1813,12 @@ mod test {
         number_of_linear_solver_setups: 23
         number_of_steps: 47
         number_of_error_test_failures: 8
-        number_of_nonlinear_solver_iterations: 119
+        number_of_nonlinear_solver_iterations: 120
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
         number_of_calls: 68
-        number_of_jac_muls: 66
+        number_of_jac_muls: 67
         number_of_matrix_evals: 3
         number_of_jac_adj_muls: 0
         "###);
@@ -1846,7 +1846,7 @@ mod test {
         number_of_linear_solver_setups: 23
         number_of_steps: 47
         number_of_error_test_failures: 8
-        number_of_nonlinear_solver_iterations: 119
+        number_of_nonlinear_solver_iterations: 120
         number_of_nonlinear_solver_fails: 0
         "###);
     }
@@ -1857,16 +1857,16 @@ mod test {
         let mut s = problem.bdf::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 75
-        number_of_steps: 319
-        number_of_error_test_failures: 5
-        number_of_nonlinear_solver_iterations: 550
[…]
