### 2026-01-06 818a395
increase default number of nls iterations
