### 2026-01-06 3238c9b
fix remaining tests
