+++ diffsol/src/ode_solver/sdirk.rs
@@ -99,9 +99,11 @@ where
             let op = op.clone_state(op.eqn().clone());
             op
         });
+        let mut convergence = self.convergence.clone();
+        convergence.reset_eta();
         Self {
             rk: self.rk.clone(),
-            convergence: self.convergence.clone(),
+            convergence,
             nonlinear_solver,
             op,
             s_op,
@@ -319,8 +321,7 @@ where
         self.update_op_step_size(h);
 
         // reinitialise jacobian updates as if a checkpoint was taken
-        self.nonlinear_solver.clear_jacobian();
-        self.convergence.reset_eta();
+        self.jacobian_updates(h, SolverState::Checkpoint);
     }
 
     fn into_state(self) -> RkState<Eqn::V> {
@@ -328,7 +329,7 @@ where
     }
 
     fn checkpoint(&mut self) -> Self::State {
-        self.nonlinear_solver.clear_jacobian();
+        self.jacobian_updates(self.rk.state().h, SolverState::Checkpoint);
         self.rk.state().clone()
     }
 
@@ -586,13 +587,13 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 8
-        number_of_steps: 32
+        number_of_steps: 31
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 119
+        number_of_nonlinear_solver_iterations: 113
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 121
+        number_of_calls: 115
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -606,14 +607,14 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 9
-        number_of_steps: 58
+        number_of_steps: 56
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 621
+        number_of_nonlinear_solver_iterations: 575
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 225
-        number_of_jac_muls: 404
+        number_of_calls: 214
+        number_of_jac_muls: 369
         number_of_matrix_evals: 2
         number_of_jac_adj_muls: 0
         "###);
@@ -628,11 +629,11 @@ mod test {
         number_of_linear_solver_setups: 7
         number_of_steps: 15
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 88
+        number_of_nonlinear_solver_iterations: 78
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 90
+        number_of_calls: 80
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -665,15 +666,15 @@ mod test {
         let mut s = problem.esdirk34_sens::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 6
[…]
