+++ diffsol/src/ode_solver/bdf.rs
@@ -2169,11 +2169,11 @@ mod test {
         );
 
         let expected_x = [
-            0.003751514915514589,
-            0.00750117409999241,
-            0.015370589755655079,
+            0.003978879413779122,
+            0.007955671343150521,
+            0.015904102550507716,
         ];
-        let expected_v = [11.202428570923361, 11.19914432101355, 11.192247396202946];
+        let expected_v = [11.202229406994425, 11.198746111635534, 11.191779520917754];
         let expected_t = [1.4281779078441663, 1.4285126937676944, 1.4292157442071036];
         for (i, ((x, v), t)) in x.iter().zip(v.iter()).zip(t.iter()).enumerate() {
             assert!(
