### 2026-01-06 59cb3b7
add min growth and max shrink back to sdirk
