+++ diffsol/src/ode_solver/sdirk.rs
@@ -581,19 +581,19 @@ mod test {
     }
 
     #[test]
-    fn test_tr_bdf2_nalgebra_exponential_decay() {
+    fn test_tr_bdf2_nalgebra_exponential_decay2() {
         let (problem, soln) = exponential_decay_problem::<M>(false);
         let mut s = problem.tr_bdf2::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 8
-        number_of_steps: 31
+        number_of_steps: 46
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 113
+        number_of_nonlinear_solver_iterations: 110
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 115
+        number_of_calls: 112
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -606,15 +606,15 @@ mod test {
         let mut s = problem.tr_bdf2_sens::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 9
-        number_of_steps: 56
+        number_of_linear_solver_setups: 10
+        number_of_steps: 74
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 575
+        number_of_nonlinear_solver_iterations: 515
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 214
-        number_of_jac_muls: 369
+        number_of_calls: 166
+        number_of_jac_muls: 357
         number_of_matrix_evals: 2
         number_of_jac_adj_muls: 0
         "###);
@@ -626,14 +626,14 @@ mod test {
         let mut s = problem.esdirk34::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 7
-        number_of_steps: 15
+        number_of_linear_solver_setups: 6
+        number_of_steps: 18
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 78
+        number_of_nonlinear_solver_iterations: 66
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 80
+        number_of_calls: 68
         number_of_jac_muls: 2
         number_of_matrix_evals: 1
         number_of_jac_adj_muls: 0
@@ -666,15 +666,15 @@ mod test {
         let mut s = problem.esdirk34_sens::<LS>().unwrap();
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
-        number_of_linear_solver_setups: 7
-        number_of_steps: 23
+        number_of_linear_solver_setups: 6
+        number_of_steps: 30
         number_of_error_test_failures: 0
-        number_of_nonlinear_solver_iterations: 364
+        number_of_nonlinear_solver_iterations: 317
         number_of_nonlinear_solver_fails: 0
         "###);
[…]
