+++ diffsol/src/ode_solver/config.rs
@@ -83,8 +83,8 @@ impl<T: Scalar> SdirkConfig<T> {
             minimum_timestep: ode_options.min_timestep,
             maximum_error_test_failures: ode_options.max_error_test_failures,
             maximum_timestep_growth: T::from_f64(2.0).unwrap(),
-            minimum_timestep_growth: T::from_f64(1.0).unwrap(),
-            maximum_timestep_shrink: T::from_f64(1.0).unwrap(),
+            minimum_timestep_growth: T::from_f64(2.0).unwrap(),
+            maximum_timestep_shrink: T::from_f64(0.9).unwrap(),
             minimum_timestep_shrink: T::from_f64(0.5).unwrap(),
             maximum_newton_iterations: ode_options.max_nonlinear_solver_iterations,
             maximum_newton_fails: ode_options.max_nonlinear_solver_failures,
