### 2026-01-06 8544d30
feat: use average convergence rate, eta persists across solves (#222)
* perf: use average rate with convergence

* fix it

* perf: remove convergence prediction

* reinstate the convergence prediction

* sdirk works again, foodweb not starting for some reason

* fix foodweb

* updated test snapshots, fixed some tests

* make ts growth/shrink more consistent

* remove rhs call in diffsl jacobian

* try tightening nls tolerance

* lets go with the tighter tolerances

* increase default number of nls iterations

* fix remaining tests

* fix ball bounce

* add min growth and max shrink back to sdi
[…]
