+++ diffsol/src/ode_solver/explicit_rk.rs
@@ -11,6 +11,7 @@ use crate::{
     AugmentedOdeEquations, DefaultDenseMatrix, DenseMatrix, ExplicitRkConfig, OdeEquations,
     OdeSolverMethod, OdeSolverProblem, OdeSolverState, Op, StateRef, StateRefMut,
 };
+use log::debug;
 use num_traits::One;
 
 impl<'a, Eqn, M, AugEqn> AugmentedOdeSolverMethod<'a, Eqn, AugEqn>
@@ -27,6 +28,9 @@ where
     }
     fn augmented_eqn(&self) -> Option<&AugEqn> {
         self.augmented_eqn.as_ref()
+    }
+    fn augmented_eqn_mut(&mut self) -> Option<&mut AugEqn> {
+        self.augmented_eqn.as_mut()
     }
 }
 
@@ -157,6 +161,11 @@ where
 
     fn step(&mut self) -> Result<OdeSolverStopReason<Eqn::T>, DiffsolError> {
         let mut h = self.rk.start_step()?;
+        debug!(
+            "Starting Explicit RK step at t = {:?} with h = {:?}",
+            self.state().t,
+            h
+        );
 
         // loop until step is accepted
         let mut nattempts = 0;
@@ -166,18 +175,27 @@ where
             for i in 1..self.rk.tableau().s() {
                 self.rk.do_stage(i, h, self.augmented_eqn.as_mut());
             }
-            let error_norm = self.rk.error_norm(h, self.augmented_eqn.as_mut());
+            let error_norm =
+                self.rk
+                    .error_norm(h, self.augmented_eqn.as_mut(), |_err: &mut Eqn::V| Ok(()))?;
             let factor = self.rk.factor(
                 error_norm,
                 1.0,
                 self.config.minimum_timestep_shrink,
+                self.config.maximum_timestep_shrink,
+                self.config.minimum_timestep_growth,
                 self.config.maximum_timestep_growth,
             );
             if error_norm < Eqn::T::one() {
+                debug!("Step accepted with error norm = {:?}", error_norm);
                 break factor;
             }
             h *= factor;
             nattempts += 1;
+            debug!(
+                "Step rejected with error norm = {:?}, reducing h to {:?}",
+                error_norm, h
+            );
             self.rk.error_test_fail(
                 h,
                 nattempts,
@@ -295,13 +313,13 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, false);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 0
-        number_of_steps: 5
+        number_of_steps: 9
         number_of_error_test_failures: 0
         number_of_nonlinear_solver_iterations: 0
         number_of_nonlinear_solver_fails: 0
         "###);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
-        number_of_calls: 32
+        number_of_calls: 56
         number_of_jac_muls: 0
         number_of_matrix_evals: 0
         number_of_jac_adj_muls: 0
@@ -326,7 +344,7 @@ mod test {
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
         number_of_linear_solver_setups: 0
         number_of_steps: 93
-        number_of_error_test_failures: 9
+        number_of_error_test_failures: 11
         number_of_nonlinear_solver_iterations: 0
         number_of_nonlinear_solver_fails: 0
         "###);
@@ -366,14 +384,14 @@ mod test {
         test_ode_solver(&mut s, soln, None, false, true);
         insta::assert_yaml_snapshot!(s.get_statistics(), @r###"
[…]
