### 2026-01-12 18effb8
fix: build pred-prey fit example, fix compile errors
