+++ examples/predator-prey-fitting-forward/Cargo.toml
@@ -5,10 +5,10 @@ edition.workspace = true
 publish = false
 
 [features]
-diffsl-llvm15 = ["diffsol/diffsl-llvm15"]
-diffsl-llvm16 = ["diffsol/diffsl-llvm16"]
-diffsl-llvm17 = ["diffsol/diffsl-llvm17"]
-diffsl-llvm18 = ["diffsol/diffsl-llvm18"]
+diffsl-llvm15 = ["diffsol/diffsl-llvm15", "diffsl-llvm"]
+diffsl-llvm16 = ["diffsol/diffsl-llvm16", "diffsl-llvm"]
+diffsl-llvm17 = ["diffsol/diffsl-llvm17", "diffsl-llvm"]
+diffsl-llvm18 = ["diffsol/diffsl-llvm18", "diffsl-llvm"]
 diffsl-llvm = []
 
 [dependencies]
