### 2026-01-12 38bd482
fix in statements in diffsl code
