+++ examples/python-diffsol/test/test_solve.py
@@ -7,8 +7,7 @@ class TestStringMethods(unittest.TestCase):
     def test_solve(self):
         model = PyDiffsol(
             """
-            in = [r, k]
-            r { 1 } k { 1 }
+            in_i { r = 1, k = 1 }
             u_i { y = 0.1 }
             F_i { (r * y) * (1 - (y / k)) }
             """
