+++ examples/population-dynamics/src/main.rs
@@ -61,8 +61,7 @@ fn phase_plane() {
         .p([1.0])
         .build_from_diffsl::<CG>(
             "
-        in = [ y0 ]
-        y0 { 1.0 }
+        in { y0 = 1.0 }
         a { 2.0/3.0 } b { 4.0/3.0 } c { 1.0 } d { 1.0 }
         u_i {
             y1 = y0,
