+++ examples/predator-prey-fitting-forward/src/main_llvm.rs
@@ -82,8 +82,8 @@ impl Gradient for Problem {
 
 pub fn main() {
     let code = "
-            in = [ b, d ]
-            a { 2.0/3.0 } b { 4.0/3.0 } c { 1.0 } d { 1.0 } x0 { 1.0 } y0 { 1.0 }
+            in_i { b = 4.0/3.0, d = 1.0 }
+            a { 2.0/3.0 } c { 1.0 } x0 { 1.0 } y0 { 1.0 }
             u_i {
                 y1 = x0,
                 y2 = y0,
