### 2026-01-12 30939b4
fix: build pred-prey fit example, fix compile errors (#225)
* fix: build pred-prey fit example, fix compile errors

* fix in statements in diffsl code
