### 2026-01-12 martinjrobins/diffsol#226 @Armavica
Missing code lines in the documentation
Hi, as part of the review that I am doing for JOSS (https://github.com/openjournals/joss-reviews/issues/9384) I am going through the diffsol book.

Thank you for fixing issue #216. I managed to run this example using the file in the repository, but because the file changed, the documentation is now missing some lines. If you copy-paste all of the codeblocks [from this page](https://martinjrobins.github.io/diffsol/primer/population_dynamics_fitting.html) you get a non-compiling file.
