### 2026-01-12 martinjrobins/diffsol#227 @Armavica
Compilation error in the neural prediction example
Hi, as part of the review that I am doing for JOSS (https://github.com/openjournals/joss-reviews/issues/9384) I am going through the diffsol book.

I am trying to run the [neural ODE weather prediction example](https://martinjrobins.github.io/diffsol/primer/weather_neural_ode.html).

These are the steps that I took:
1. I am using the directory `diffsol/examples/neural-ode-weather-prediction`
2. I initialized a Python environment with the packages specified in the `requirements.txt` file
3. I ran the command `cargo run --release --features="onnx"`

The compilation fails with the message below. Is there something that I am doing incorrectly?

<details><summary>Compilation log</summary>

```
❯ cargo run --release --features="onnx"
   Compiling proc-macro2 v1.0.103
   Compiling unicode-ident v1.0.22
   Compiling quote v1.0.42
   Compiling autocfg v1.5.0
   Compiling libm v0.2.15
   Compiling libc v0.2.178
   Compiling bytemuck v1.24.0
   Compiling version_check v0.9.5
   Compiling rand_core v0.6.4
   Compiling crossbeam-utils v0.8.21
   Compiling cfg-if v1.0.4
   Compiling stable_deref_trait v1.2.1
   Compiling typenum v1.19.0
   Compiling nano-gemm-codegen v0.1.0
   Compiling serde_core v1.0.228
   Compiling rayon-core v1.13.0
   Compiling serde v1.0.228
   Compiling paste v1.0.15
   Compiling memchr v2.7.6
   Compiling rand v0.8.5
   Compiling shlex v1.3.0
   Compiling litemap v0.8.1
   Compiling find-msvc-tools v0.1.5
   Compiling once_cell v1.21.3
   Compiling writeable v0.6.2
   Compiling pulp v0.21.5
   Compiling num-traits v0.2.19
   Compiling cc v1.2.49
   Compiling icu_properties_data v2.1.2
   Compiling reborrow v0.5.5
   Compiling zerocopy v0.8.31
   Compiling either v1.15.0
   Compiling icu_normalizer_data v2.1.1
   Compiling generic-array v0.14.7
   Compiling bitflags v2.10.0
   Compiling zeroize v1.8.2
   Compiling dyn-stack-macros v0.1.3
   Compiling crossbeam-epoch v0.9.18
   Compiling rustls-pki-types v1.13.2
   Compiling raw-cpuid v11.6.0
   Compiling dyn-stack v0.13.2
   Compiling syn v2.0.111
   Compiling smallvec v1.15.1
   Compiling crossbeam-deque v0.8.6
   Compiling ucd-trie v0.1.7
   Compiling getrandom v0.3.4
   Compiling pest v2.8.4
   Compiling untrusted v0.9.0
   Compiling rustix v1.1.2
   Compiling seq-macro v0.3.6
   Compiling syn v1.0.109
   Compiling getrandom v0.2.16
   Compiling matrixmultiply v0.3.10
   Compiling linux-raw-sys v0.11.0
   Compiling crc32fast v1.5.0
   Compiling rustls v0.23.35
   Compiling fnv v1.0.7
   Compiling percent-encoding v2.3.2
   Compiling utf8_iter v1.0.4
   Compiling nano-gemm-core v0.1.0
   Compiling ident_case v1.0.1
   Compiling strsim v0.11.1
   Compiling form_urlencoded v1.2.2
   Compiling rayon v1.11.0
   Compiling rand_core v0.9.3
   Compiling ring v0.17.14
   Compiling webpki-roots v1.0.4
   Compiling num-complex v0.4.6
   Compiling num-integer v0.1.46
   Compiling crypto-common v0.1.7
   Compiling block-buffer v0.10.4
   Compiling rawpointer v0.2.1
   Compiling subtle v2.6.1
   Compiling byteorder v1.5.0
   Compiling log v0.4.29
   Compiling simd-adler32 v0.3.8
   Compiling pest_meta v2.8.4
   Compiling adler2 v2.0.1
   Compiling socks v0.3.4
   Compiling miniz_oxide v0.8.9
   Compiling digest v0.10.7
   Compiling webpki-roots v0.26.11
   Compiling filetime v0.2.26
   Compiling nano-gemm-c64 v0.1.0
   Compiling nano-gemm-f32 v0.1.0
   Compiling nano-gemm-f64 v0.1.0
   Compiling nano-gemm-c32 v0.1.0
   Compiling crossbeam-channel v0.5.15
   Compiling crossbeam-queue v0.3.12
   Compiling safe_arch v0.7.4
   Compiling ryu v1.0.20
   Compiling serde_json v1.0.145
   Compiling defer v0.2.1
   Compiling xattr v1.6.1
   Compiling cpufeatures v0.2.17
   Compiling base64 v0.22.1
   Compiling itoa v1.0.15
   Compiling crossbeam v0.8.4
   Compiling wide v0.7.33
   Compiling tar v0.4.44
   Compiling sha2 v0.10.9
   Compiling flate2 v1.1.5
   Compiling num-bigint v0.4.6
   Compiling approx v0.5.1
   Compiling atomic-wait v1.1.0
   Compiling winnow v0.7.14
   Compiling typeid v1.0.3
   Compiling pkg-config v0.3.32
   Compiling num-rational v0.4.2
   Compiling num_cpus v1.17.0
   Compiling rustc-hash v2.1.1
   Compiling equivalent v1.0.2
   Compiling synstructure v0.13.2
   Compiling gemm-common v0.18.2
   Compiling pest_generator v2.8.4
   Compiling darling_core v0.21.3
   Compiling rustls-webpki v0.103.8
   Compiling interpol v0.2.1
   Compiling gemm-f64 v0.18.2
   Compiling private-gemm-x86 v0.1.20
   Compiling gemm-c64 v0.18.2
   Compiling gemm-c32 v0.18.2
   Compiling gemm-f32 v0.18.2
   Compiling zerofrom-derive v0.1.6
   Compiling yoke-derive v0.8.1
   Compiling zerovec-derive v0.11.2
   Compiling displaydoc v0.2.5
   Compiling zerofrom v0.1.6
   Compiling serde_derive v1.0.228
   Compiling yoke v0.8.1
   Compiling zerocopy-derive v0.8.31
   Compiling zerovec v0.11.5
   Compiling zerotrie v0.2.3
   Compiling pest_derive v2.8.4
   Compiling equator-macro v0.4.2
   Compiling equator-macro v0.2.1
   Compiling tinystr v0.8.2
   Compiling potential_utf v0.1.4
   Compiling icu_collections v2.1.1
   Compiling icu_locale_core v2.1.1
   Compiling darling_macro v0.21.3
   Compiling equator v0.4.2
   Compiling equator v0.2.2
   Compiling py_literal v0.4.0
   Compiling darling v0.21.3
   Compiling spindle v0.2.6
   Compiling nalgebra-macros v0.3.0
   Compiling faer-macros v0.22.1
   Compiling simba v0.9.1
   Compiling icu_provider v2.1.1
   Compiling qd v0.7.7
   Compiling foldhash v0.1.5
   Compiling generativity v1.1.0
   Compiling thiserror v2.0.17
   Compiling icu_normalizer v2.1.1
   Compiling icu_properties v2.1.2
   Compiling hashbrown v0.16.1
   Compiling erased-serde v0.4.9
   Compiling npyz v0.8.4
   Compiling hashbrown v0.15.5
   Compiling faer-traits v0.23.2
   Compiling indexmap v2.12.1
   Compiling idna_adapter v1.2.1
   Compiling idna v1.1.0
   Compiling ppv-lite86 v0.2.21
   Compiling serde_with_macros v3.16.1
   Compiling gemm v0.18.2
   Compiling thiserror-impl v2.0.17
   Compiling basic-toml v0.1.10
   Compiling rand_chacha v0.9.0
   Compiling url v2.5.7
   Compiling askama_parser v0.14.0
   Compiling rand v0.9.2
   Compiling tracing-core v0.1.35
   Compiling fixedbitset v0.5.7
   Compiling pin-project-lite v0.2.16
   Compiling diffsol v0.10.0 (/home/virgile/src/diffsol/diffsol)
   Compiling serde_with v3.16.1
   Compiling ureq v2.12.1
   Compiling tracing v0.1.43
   Compiling rand_distr v0.5.1
   Compiling petgraph v0.8.3
   Compiling half v2.7.1
   Compiling plotly_derive v0.13.5
   Compiling serde_repr v0.1.20
   Compiling ndarray v0.16.1
   Compiling csv-core v0.1.13
   Compiling dyn-clone v1.0.20
   Compiling csv v1.4.0
   Compiling ort-sys v2.0.0-rc.9
   Compiling nalgebra v0.34.1
   Compiling askama_derive v0.14.0
   Compiling nano-gemm v0.1.3
   Compiling faer v0.23.2
   Compiling askama v0.14.0
   Compiling plotly v0.13.5
   Compiling ort v2.0.0-rc.9
   Compiling nalgebra-sparse v0.11.0
   Compiling neural-ode-weather-prediction v0.1.0 (/home/virgile/src/diffsol/examples/neural-ode-weather-prediction)
error: linking with `cc` failed: exit status: 1
  |
  = note:  "cc" "-m64" "/tmp/rustcdLdCNS/symbols.o" "<17 object files omitted>" "-Wl,--as-needed" "-Wl,-Bstatic" "/home/virgile/src/diffsol/target/release/deps/{libplotly-b49503b84ec963ff,libserde_with-089be1704d38b66b,liberased_serde-7dcce72e04aa62d7,libtypeid-8703836c76c84d06,libdyn_clone-af5d4b7c202cb2a0,libaskama-586da58daadbd4e1,libpercent_encoding-4c765ea3fe4e563c,libort-dc64e4a238ddd1c5,libhalf-f398c5cdb7161e24,libtracing-5c291444f76f443a,libpin_project_lite-56f53742df337c04,libtracing_core-7ddb4a23ab556c8a,libort_sys-fb33180b60b08a86,libndarray-dfd07b9b1460f0d5,libdiffsol-ad708a1f7a29ff26,libserde_json-b000b2f1da9b0169,libthiserror-ec7e4be57edf855c,libserde-eb15ff00fd081725,liblog-4fe4b0f6afd99737,libnalgebra-dda83d08b45e7bc2,libmatrixmultiply-194309a9d0c77194,librawpointer-aa726d2013a1c163,libnum_rational-eac17e2b9604009e,libtypenum-3aaece5f1769cdc7,libsimba-4c3147ce9c9177ef,libwide-1f7524ccc591ce81,libsafe_arch-337940eacea15b38,libapprox-3b9caebaab4b493b,libpetgraph-f139531e6ff4cdda,libhashbrown-163d7acfd22aae69,libfoldhash-c9861e4771522c87,libindexmap-53258b48d4c81dce,libequivalent-cae8259c6c10db17,libhashbrown-b59f6065a9008371,libfixedbitset-a60c231fdb5c3572,libfaer-3a8208b7a14eba22,libnpyz-2dd34595f3f1b4d9,libbyteorder-40b5924803c7721a,libpy_literal-df7804016fab45f8,libpest-f99632e010b9fe96,libucd_trie-e74b077d3b298ba6,libnum_bigint-507d37067c90f54e,libnum_integer-89bac257ee6a0ad9,libgemm-03111f6fe0f6f389,libgemm_c32-2eccc79be1ea8bec,libgemm_c64-7bee9e72a03692f0,libgemm_f32-aa91c5b8a0b7ce1f,libgemm_f64-1c81663e76e6c885,libgemm_common-556c6bdf00733e68,libonce_cell-dc7066c855677814,libnano_gemm-47142498ab7816d1,libnano_gemm_f64-c1405094541e6b96,libnano_gemm_f32-7fff537f7535740a,libnano_gemm_c64-c8392fe9adc788ee,libnano_gemm_c32-15fb894b3d3ba426,libnano_gemm_core-ffcc87457424e6d5,libequator-b4fe74f1c184d615,libprivate_gemm_x86-c6fed340e9d9a6e6,libspindle-108dbe4390b329c7,libatomic_wait-d53e4cbb2695cad5,libnum_cpus-17723b64c601bb74,libraw_cpuid-00f3a4b28b42c00d,libbitflags-ed649bd01a93bc44,librand_distr-db19a5722b6ab278,librayon-22393e1388b09e66,librayon_core-5cb507c08d904a2f,libcrossbeam_deque-abaa8104975aeba9,libcrossbeam_epoch-647f54b08ada27e0,libcrossbeam_utils-68fd8b948f72a483,libeither-761755319f95a234,libgenerativity-c639f4c5e69c579d,libequator-366f6a6dd2666f0d,librand-d99a90b8ea7429b3,librand_chacha-ff574af526ad9313,libppv_lite86-c707a85e0723a7ac,libzerocopy-36fbf4ec6e08fa70,librand_core-7e8aec74b79b2394,libgetrandom-913291bab1917344,liblibc-a58828e456c21036,libcfg_if-ebc73076c81e0e49,libreborrow-17e671d31196489f,libfaer_traits-c39bbcc52e063cf8,liblibm-dc80b440da584766,libqd-f76351c2b84dec95,libpulp-e7b3c1a3c1d46a66,libnum_complex-2dfeada4f0f370f1,librand-9b097f82dbcdaec9,librand_core-f0ee8e9803370321,libnum_traits-ba6dce69ad501a5e,libdyn_stack-0b70b53889f41b96,libbytemuck-3f7f36c8c0c83591,libcsv-e4a6551fa0a9bb43,libryu-10f98fa487b660e3,libitoa-55109e383f544441,libcsv_core-90230f1f628d5265,libmemchr-d87f7b66dbd7bd49,libserde_core-d3b5aef1a7adabc2}.rlib" "<sysroot>/lib/rustlib/x86_64-unknown-linux-gnu/lib/{libstd-*,libpanic_unwind-*,libobject-*,libmemchr-*,libaddr2line-*,libgimli-*,libcfg_if-*,librustc_demangle-*,libstd_detect-*,libhashbrown-*,librustc_std_workspace_alloc-*,libminiz_oxide-*,libadler2-*,libunwind-*,liblibc-*,librustc_std_workspace_core-*,liballoc-*,libcore-*,libcompiler_builtins-*}.rlib" "-Wl,-Bdynamic" "-lstdc++" "-lgcc_s" "-lutil" "-lrt" "-lpthread" "-lm" "-ldl" "-lc" "-L" "/tmp/rustcdLdCNS/raw-dylibs" "-B<sysroot>/lib/rustlib/x86_64-unknown-linux-gnu/bin/gcc-ld" "-fuse-ld=lld" "-Wl,--eh-frame-hdr" "-Wl,-z,noexecstack" "-L" "/home/virgile/.cache/ort.pyke.io/dfbin/x86_64-unknown-linux-gnu/F88A8C1E4B4813A1CFA79AF3F35B23ADDF2F0F36E66C5CD7C88103CB9B30509D/onnxruntime/lib" "-L" "<sysroot>/lib/rustlib/x86_64-unknown-linux-gnu/lib" "-o" "/home/virgile/src/diffsol/target/release/deps/neural_ode_weather_prediction-9bd9fe01debf8532" "-Wl,--gc-sections" "-pie" "-Wl,-z,relro,-z,now" "-Wl,-O1" "-Wl,--strip-debug" "-nodefaultlibs"
  = note: some arguments are omitted. use `--verbose` to show all linker arguments
  = note: rust-lld: error: undefined symbol: OrtGetApiBase
          >>> referenced by ort.2aedd0ff21153b7b-cgu.10
          >>>               ort-dc64e4a238ddd1c5.ort.2aedd0ff21153b7b-cgu.10.rcgu.o:(std::sync::once::Once::call_once_force::_$u7b$$u7b$closure$u7d$$u7d$::h65c230b753f91752 (.llvm.17207353300941267976)) in archive /home/virgile/src/diffsol/target/release/deps/libort-dc64e4a238ddd1c5.rlib
          >>> referenced by ort.2aedd0ff21153b7b-cgu.10
          >>>               ort-dc64e4a238ddd1c5.ort.2aedd0ff21153b7b-cgu.10.rcgu.o:(core::ops::function::FnOnce::call_once$u7b$$u7b$vtable.shim$u7d$$u7d$::h108721a84277271a (.llvm.17207353300941267976)) in archive /home/virgile/src/diffsol/target/release/deps/libort-dc64e4a238ddd1c5.rlib
          collect2: error: ld returned 1 exit status

error: could not compile `neural-ode-weather-prediction` (bin "neural-ode-weather-prediction") due to 1 previous error

```
