### 2026-01-13 c3766ed
doc: add some api docs for problem struct
