### 2026-01-13 0aa8412
finish adding docstring for problem
