+++ diffsol/src/ode_solver/problem.rs
@@ -123,6 +123,14 @@ where
 
 macro_rules! sdirk_solver_from_tableau {
     ($method:ident, $method_sens:ident, $method_solver:ident, $method_solver_sens:ident, $method_solver_adjoint:ident, $tableau:ident) => {
+        #[doc = concat!("Create a new ", stringify!($tableau), " SDIRK solver instance with the given initial state.\n\n",
+            "This method uses the built-in ", stringify!($tableau), " Butcher tableau.\n\n",
+            "# Type Parameters\n",
+            "- `LS`: The linear solver type\n\n",
+            "# Arguments\n",
+            "- `state`: The initial state for the solver\n\n",
+            "# Returns\n",
+            "An SDIRK solver instance configured with the ", stringify!($tableau), " method")]
         pub fn $method_solver<LS: LinearSolver<Eqn::M>>(
             &self,
             state: RkState<Eqn::V>,
@@ -136,6 +144,15 @@ macro_rules! sdirk_solver_from_tableau {
             )
         }
 
+        #[doc = concat!("Create a new ", stringify!($tableau), " SDIRK solver instance with forward sensitivities, given the initial state.\n\n",
+            "This method uses the built-in ", stringify!($tableau), " Butcher tableau and simultaneously solves\n",
+            "the state equations and forward sensitivity equations.\n\n",
+            "# Type Parameters\n",
+            "- `LS`: The linear solver type\n\n",
+            "# Arguments\n",
+            "- `state`: The initial state for the solver (including sensitivities)\n\n",
+            "# Returns\n",
+            "An SDIRK solver instance configured for forward sensitivity analysis using ", stringify!($tableau))]
         pub fn $method_solver_sens<LS: LinearSolver<Eqn::M>>(
             &self,
             state: RkState<Eqn::V>,
@@ -152,6 +169,17 @@ macro_rules! sdirk_solver_from_tableau {
             )
         }
 
+        #[doc = concat!("Create a new ", stringify!($tableau), " SDIRK solver instance for adjoint sensitivity analysis.\n\n",
+            "This method creates a solver for the backward adjoint equations using the ", stringify!($tableau), " method.\n",
+            "Requires a checkpointer to provide the forward solution during the backward solve.\n\n",
+            "# Type Parameters\n",
+            "- `LS`: The linear solver type\n",
+            "- `S`: The forward solver method type used for checkpointing\n\n",
+            "# Arguments\n",
+            "- `checkpointer`: The checkpointing object containing the forward solution\n",
+            "- `nout_override`: Optional override for the number of output equations\n\n",
+            "# Returns\n",
+            "An SDIRK solver instance configured for adjoint sensitivity analysis using ", stringify!($tableau))]
         pub fn $method_solver_adjoint<'a, LS: LinearSolver<Eqn::M>, S: OdeSolverMethod<'a, Eqn>>(
             &'a self,
             checkpointer: Checkpointing<'a, Eqn, S>,
@@ -170,6 +198,14 @@ macro_rules! sdirk_solver_from_tableau {
             )
         }
 
+        #[doc = concat!("Create a new ", stringify!($tableau), " SDIRK solver instance with a consistent initial state.\n\n",
+            "This convenience method combines state creation and solver initialization using the\n",
+            "built-in ", stringify!($tableau), " Butcher tableau. It will create a consistent initial state,\n",
+            "which may require solving a nonlinear system if a mass matrix is present.\n\n",
+            "# Type Parameters\n",
+            "- `LS`: The linear solver type\n\n",
+            "# Returns\n",
[…]
