### 2026-01-13 16b4477
finish more docstrings
