### 2026-01-13 e87f71e
doc: fill out documentation for builder
