+++ diffsol/src/ode_solver/builder.rs
@@ -274,6 +274,16 @@ where
         }
     }
 
+    /// Set the right-hand side of the ODE for adjoint sensitivity analysis.
+    ///
+    /// This method provides all the functions needed for computing adjoint sensitivities,
+    /// including the transpose Jacobian operations.
+    ///
+    /// # Arguments
+    /// - `rhs`: Function of type Fn(x: &V, p: &V, t: S, y: &mut V) that computes the right-hand side of the ODE.
+    /// - `rhs_jac`: Function of type Fn(x: &V, p: &V, t: S, v: &V, y: &mut V) that computes the multiplication of the Jacobian of the right-hand side with the vector v.
+    /// - `rhs_adjoint`: Function of type Fn(x: &V, p: &V, t: S, v: &V, y: &mut V) that computes the multiplication of the transpose of the Jacobian with the vector v.
+    /// - `rhs_sens_adjoint`: Function of type Fn(x: &V, p: &V, t: S, v: &V, y: &mut V) that computes the multiplication of the transpose of the partial derivative of the rhs wrt the parameters, with the vector v.
     #[allow(clippy::type_complexity)]
     pub fn rhs_adjoint_implicit<F, G, H, I>(
         self,
@@ -595,6 +605,15 @@ where
         }
     }
 
+    /// Set the output function of the ODE.
+    ///
+    /// Output functions compute additional quantities of interest that depend on the state.
+    /// These can be integrated alongside the ODE or used for root finding.
+    ///
+    /// # Arguments
+    /// - `out`: Function of type Fn(x: &V, p: &V, t: S, y: &mut V) that computes the output.
+    /// - `out_jac`: Function of type Fn(x: &V, p: &V, t: S, v: &V, y: &mut V) that computes the multiplication of the Jacobian of the output with the vector v.
+    /// - `nout`: Number of output equations (i.e. size of the output vector `y`).
     pub fn out_implicit<F, G>(
         self,
         out: F,
@@ -638,6 +657,16 @@ where
         }
     }
 
+    /// Set the output function of the ODE for forward sensitivity analysis, suitable for implicit solvers.
+    ///
+    /// This extends the output function with sensitivity information needed for
+    /// computing sensitivities of the outputs with respect to parameters.
+    ///
+    /// # Arguments
+    /// - `out`: Function of type Fn(x: &V, p: &V, t: S, y: &mut V) that computes the output.
+    /// - `out_jac`: Function of type Fn(x: &V, p: &V, t: S, v: &V, y: &mut V) that computes the multiplication of the Jacobian of the output with the vector v.
+    /// - `out_sens`: Function of type Fn(x: &V, p: &V, t: S, v: &V, y: &mut V) that computes the multiplication of the partial derivative of the output wrt the parameters, with the vector v.
+    /// - `nout`: Number of output equations (i.e. size of the output vector `y`).
     pub fn out_sens_implicit<F, G, H>(
         self,
         out: F,
@@ -684,6 +713,17 @@ where
         }
     }
 
+    /// Set the output function of the ODE for adjoint sensitivity analysis, suitable for implicit solvers.
+    ///
+    /// This provides all the functions needed for computing adjoint sensitivities of outputs,
+    /// including the transpose Jacobian operations.
+    ///
+    /// # Arguments
+    /// - `out`: Function of type Fn(x: &V, p: &V, t: S, y: &mut V) that computes the output.
+    /// - `out_jac`: Function of type Fn(x: &V, p: &V, t: S, v: &V, y: &mut V) that computes the multiplication of the Jacobian of the output with the vector v.
+    /// - `out_adjoint`: Function of type Fn(x: &V, p: &V, t: S, v: &V, y: &mut V) that computes the multiplication of the transpose of the Jacobian with the vector v.
[…]
