### 2026-01-13 1771b99
doc: checkpointing, state
