### 2026-01-13 ab2c2ea
doc: improve OdeSolverMethod
