### 2026-01-13 d8e3fb9
doc: improve Vector
