+++ diffsol/src/vector/mod.rs
@@ -17,6 +17,10 @@ pub mod cuda;
 #[macro_use]
 mod utils;
 
+/// A trait for types that represent a collection of indices into a vector.
+///
+/// This is used to represent subsets of vector elements, typically for algebraic constraints
+/// or when operating on specific vector components.
 pub trait VectorIndex: Sized + Debug + Clone {
     type C: Context;
     fn context(&self) -> &Self::C;
@@ -29,6 +33,7 @@ pub trait VectorIndex: Sized + Debug + Clone {
     fn from_vec(v: Vec<IndexType>, ctx: Self::C) -> Self;
 }
 
+/// Common interface for vector-like types, providing access to scalar type, context, and inner representation.
 pub trait VectorCommon: Sized + Debug {
     type T: Scalar;
     type C: Context;
@@ -61,6 +66,9 @@ where
     }
 }
 
+/// Operations on vectors by value (addition and subtraction).
+///
+/// This trait defines vector addition and subtraction when both operands are owned or references.
 pub trait VectorOpsByValue<Rhs = Self, Output = Self>:
     VectorCommon + Add<Rhs, Output = Output> + Sub<Rhs, Output = Output>
 {
@@ -71,10 +79,17 @@ impl<V, Rhs, Output> VectorOpsByValue<Rhs, Output> for V where
 {
 }
 
+/// In-place operations on vectors (addition and subtraction).
+///
+/// This trait defines in-place vector addition and subtraction (self += rhs, self -= rhs).
 pub trait VectorMutOpsByValue<Rhs = Self>: VectorCommon + AddAssign<Rhs> + SubAssign<Rhs> {}
 
 impl<V, Rhs> VectorMutOpsByValue<Rhs> for V where V: VectorCommon + AddAssign<Rhs> + SubAssign<Rhs> {}
 
+/// Operations on a reference to a vector, supporting addition, subtraction, and scalar multiplication.
+///
+/// This trait ensures that vector references can be used in arithmetic operations with owned vectors,
+/// other references, and vector views, enabling flexible composition of vector operations.
 pub trait VectorRef<V: Vector>:
     VectorOpsByValue<V, V>
     + for<'a> VectorOpsByValue<&'a V, V>
@@ -93,6 +108,11 @@ impl<RefT, V: Vector> VectorRef<V> for RefT where
 {
 }
 
+/// A mutable view into a vector, supporting in-place operations and modifications.
+///
+/// This trait represents a temporary mutable reference to a vector's data, allowing in-place
+/// arithmetic operations (+=, -=, *=) and other modifications. Mutable views can be created
+/// via the `as_view_mut()` method on a `Vector`.
 pub trait VectorViewMut<'a>:
     VectorMutOpsByValue<Self::View>
     + VectorMutOpsByValue<Self::Owned>
@@ -103,11 +123,19 @@ pub trait VectorViewMut<'a>:
     type Owned;
     type View;
     type Index: VectorIndex;
+    /// Copy values from an owned vector into this view.
     fn copy_from(&mut self, other: &Self::Owned);
+    /// Copy values from another vector view into this view.
     fn copy_from_view(&mut self, other: &Self::View);
+    /// Compute the AXPY operation: self = alpha * x + beta * self
     fn axpy(&mut self, alpha: Self::T, x: &Self::Owned, beta: Self::T);
 }
 
+/// A borrowed immutable view of a vector, supporting read-only arithmetic operations.
+///
+/// This trait represents a temporary immutable reference to a vector's data, allowing read-only
+/// operations like addition, subtraction, and scalar multiplication. Vector views can be created
+/// via the `as_view()` method on a `Vector` and are cheaper to create than cloning.
 pub trait VectorView<'a>:
     VectorOpsByValue<Self, Self::Owned>
[…]
