### 2026-01-13 52f247c
doc: improve Matrix
