+++ diffsol/src/matrix/mod.rs
@@ -29,14 +29,18 @@ pub mod sparsity;
 #[macro_use]
 mod utils;
 
+/// Common interface for matrix types, providing access to scalar type, context, and dimensions.
 pub trait MatrixCommon: Sized + Debug {
     type V: Vector<T = Self::T, C = Self::C, Index: VectorIndex<C = Self::C>>;
     type T: Scalar;
     type C: Context;
     type Inner;
 
+    /// Get the number of rows in this matrix.
     fn nrows(&self) -> IndexType;
+    /// Get the number of columns in this matrix.
     fn ncols(&self) -> IndexType;
+    /// Get a reference to the inner representation of the matrix.
     fn inner(&self) -> &Self::Inner;
 }
 
@@ -80,6 +84,9 @@ where
     }
 }
 
+/// Operations on matrices by value (addition and subtraction).
+///
+/// This trait defines matrix addition and subtraction when both operands are owned or references.
 pub trait MatrixOpsByValue<Rhs = Self, Output = Self>:
     MatrixCommon + Add<Rhs, Output = Output> + Sub<Rhs, Output = Output>
 {
@@ -90,6 +97,9 @@ impl<M, Rhs, Output> MatrixOpsByValue<Rhs, Output> for M where
 {
 }
 
+/// In-place operations on matrices (addition and subtraction).
+///
+/// This trait defines in-place matrix addition and subtraction (self += rhs, self -= rhs).
 pub trait MatrixMutOpsByValue<Rhs = Self>: MatrixCommon + AddAssign<Rhs> + SubAssign<Rhs> {}
 
 impl<M, Rhs> MatrixMutOpsByValue<Rhs> for M where M: MatrixCommon + AddAssign<Rhs> + SubAssign<Rhs> {}
@@ -98,7 +108,11 @@ impl<M, Rhs> MatrixMutOpsByValue<Rhs> for M where M: MatrixCommon + AddAssign<Rh
 pub trait MatrixRef<M: MatrixCommon>: Mul<Scale<M::T>, Output = M> {}
 impl<RefT, M: MatrixCommon> MatrixRef<M> for RefT where RefT: Mul<Scale<M::T>, Output = M> {}
 
-/// A mutable view of a dense matrix [Matrix]
+/// A mutable view of a dense matrix, supporting in-place operations and modifications.
+///
+/// This trait represents a temporary mutable reference to a matrix's data, allowing in-place
+/// arithmetic operations (+=, -=, *=) and matrix-matrix multiplication. Mutable views can be
+/// created via the `columns_mut()` or `column_mut()` methods on a `DenseMatrix`.
 pub trait MatrixViewMut<'a>:
     for<'b> MatrixMutOpsByValue<&'b Self>
     + for<'b> MatrixMutOpsByValue<&'b Self::View>
@@ -106,20 +120,28 @@ pub trait MatrixViewMut<'a>:
 {
     type Owned;
     type View;
+    /// Convert this mutable view into an owned matrix, cloning the data if necessary.
     fn into_owned(self) -> Self::Owned;
+    /// Perform matrix-matrix multiplication with owned matrices: self = alpha * a * b + beta * self
     fn gemm_oo(&mut self, alpha: Self::T, a: &Self::Owned, b: &Self::Owned, beta: Self::T);
+    /// Perform matrix-matrix multiplication with a view and owned matrix: self = alpha * a * b + beta * self
     fn gemm_vo(&mut self, alpha: Self::T, a: &Self::View, b: &Self::Owned, beta: Self::T);
 }
 
-/// A view of a dense matrix [Matrix]
+/// A borrowed immutable view of a dense matrix, supporting read-only arithmetic operations.
+///
[…]
