+++ diffsol/src/ode_solver/problem.rs
@@ -84,7 +84,7 @@ impl<T: Scalar> Default for OdeSolverOptions<T> {
 /// tolerances, initial conditions, and solver options. This struct can be used
 /// to create individual solvers for different methods (e.g., BDF, Runge-Kutta),
 /// solving the same underlying problem with consistent settings.
-/// 
+///
 /// This struct is normally generated via the `OdeBuilder` API, which provides
 /// a more user-friendly interface for constructing ODE solver problems.
 pub struct OdeSolverProblem<Eqn>
@@ -355,7 +355,7 @@ where
     Eqn: OdeEquations,
 {
     /// Returns whether outputs are included in the error control.
-    /// 
+    ///
     /// This returns `true` if all of the following conditions are met:
     /// - Output integration is enabled (`integrate_out` is true)
     /// - The equations have output functions defined
@@ -442,17 +442,17 @@ where
     }
 
     /// Create a new BDF solver instance for the problem with the given initial state.
-    /// 
+    ///
     /// This method creates a BDF solver with a Newton nonlinear solver using the specified
     /// linear solver type. The state must be provided and should typically be created using
     /// [`Self::bdf_state`], which ensures consistency for problems with mass matrices.
-    /// 
+    ///
     /// # Type Parameters
     /// - `LS`: The linear solver type used in the Newton solver
-    /// 
+    ///
     /// # Arguments
     /// - `state`: The initial state for the solver
-    /// 
+    ///
     /// # Returns
     /// A BDF solver instance configured for this problem
     #[allow(clippy::type_complexity)]
@@ -469,12 +469,12 @@ where
 
     /// Create a new BDF solver instance for the problem. This will create a consistent initial state,
     /// so might require solving a nonlinear system if a mass matrix is present.
-    /// 
+    ///
     /// This is a convenience method that combines [`Self::bdf_state`] and [`Self::bdf_solver`].
-    /// 
+    ///
     /// # Type Parameters
     /// - `LS`: The linear solver type used in the Newton solver
-    /// 
+    ///
     /// # Returns
     /// A BDF solver instance configured for this problem with a consistent initial state
     #[allow(clippy::type_complexity)]
@@ -514,20 +514,20 @@ where
     }
 
     /// Create a new BDF solver instance for the backwards solve for the adjoint equations. This requires
-    /// a checkpointer to provide the forward solution during the backward solve. 
-    /// 
+    /// a checkpointer to provide the forward solution during the backward solve.
+    ///
     /// If you are computing adjoint sensitivites of the continuous integral of the outputs, the number
     /// of output equations are taken from the equations being solved. If you are computing adjoint sensitivities
     /// for a discrete sum of outputs not in your equations, you must override the number of outputs via `nout_override`.
-    /// 
+    ///
     /// # Type Parameters
     /// - `LS`: The linear solver type used in the Newton solver
     /// - `S`: The forward solver method type used for checkpointing (this can be auto-deduced from the `checkpointer`)
-    /// 
+    ///
     /// # Arguments
     /// - `checkpointer`: The checkpointing object containing the forward solution
     /// - `nout_override`: Optional override for the number of output equations
-    /// 
+    ///
     /// # Returns
     /// A BDF solver instance configured for adjoint sensitivity analysis
     #[allow(clippy::type_complexity)]
@@ -572,16 +572,16 @@ where
     }
 
     /// Create a new BDF solver instance for the problem with forward sensitivities, given the initial state.
-    /// 
+    ///
[…]
