### 2026-01-13 b001b43
doc: diffsl, ops and jacobian
