+++ .github/workflows/rust.yml
@@ -148,6 +148,9 @@ jobs:
       if: matrix.os == 'windows-latest'
       run: |
         echo "ADDITIONAL_FEATURES_FLAGS=--features diffsl-cranelift" >> $GITHUB_ENV
+    - name: Free Disk Space (Ubuntu)
+      if: matrix.os == 'ubuntu-latest'
+      uses: jlumbroso/free-disk-space@main
     - name: Run tests - default features
       if: matrix.tests == true
       run: cargo test --verbose
