+++ .github/workflows/rust.yml
@@ -151,6 +151,13 @@ jobs:
     - name: Free Disk Space (Ubuntu)
       if: matrix.os == 'ubuntu-latest'
       uses: jlumbroso/free-disk-space@main
+      with:
+        android: true
+        dotnet: true
+        haskell: true
+        large-packages: false
+        docker-images: true
+        swap-storage: true
     - name: Run tests - default features
       if: matrix.tests == true
       run: cargo test --verbose
