### 2026-01-13 76b9672
doc: improve API docs (#229)
* doc: add some api docs for problem struct

* finish adding docstring for problem

* finish more docstrings

* doc: fill out documentation for builder

* doc: checkpointing, state

* doc: improve OdeSolverMethod

* doc: improve Vector

* doc: improve Matrix

* fix combine

* doc: improve Scalar

* remove some commented out code

* cargo fmt

* doc: diffsl, ops and jacobian

* cargo fmt

* doc: line_search

* doc: modules

* fix some bad links

* free up some disk space on ubuntu

* free disk space
