### 2026-01-19 a18916c
fix: solve and solve_dense respect roots in problem
