### 2026-01-19 88c5654
fix: adapt onnx example for ort rc.10 (#232)
