### 2026-01-19 6579ee5
test: assert root solution and integral
