### 2026-01-19 40e2c27
Add convergence constructor for configurable tolerance
