### 2026-01-19 d2c35df
fix: solve and solve_dense respect roots in problem (#231)
* fix: solve and solve_dense respect roots in problem

* cargo fmt

* test: assert root solution and integral

* cargo fmt

* fix clippy
