+++ diffsol/src/ode_solver/method.rs
@@ -161,6 +161,8 @@ where
     ///
     /// # Post-condition
     /// After the solver finishes, the internal state of the solver is at time `final_time`.
+    /// If a root is found, the solver stops early. The internal state is moved to the root time,
+    /// and the root time/value are returned as the last entry.
     #[allow(clippy::type_complexity)]
     fn solve(
         &mut self,
@@ -176,12 +178,39 @@ where
         // do the main loop
         write_out(self, &mut ret_y, &mut ret_t, &mut tmp_nout);
         self.set_stop_time(final_time)?;
-        while self.step()? != OdeSolverStopReason::TstopReached {
-            write_out(self, &mut ret_y, &mut ret_t, &mut tmp_nout);
+        loop {
+            match self.step()? {
+                OdeSolverStopReason::InternalTimestep => {
+                    write_out(self, &mut ret_y, &mut ret_t, &mut tmp_nout);
+                }
+                OdeSolverStopReason::TstopReached => {
+                    write_out(self, &mut ret_y, &mut ret_t, &mut tmp_nout);
+                    break;
+                }
+                OdeSolverStopReason::RootFound(t_root) => {
+                    let nstates = self.problem().eqn.rhs().nstates();
+                    let mut y_root = Eqn::V::zeros(nstates, self.problem().context().clone());
+                    self.interpolate_inplace(t_root, &mut y_root)?;
+                    let integrate_out = self.problem().integrate_out;
+                    let mut g_root = None;
+                    if integrate_out {
+                        let mut g = self.state().g.clone();
+                        self.interpolate_out_inplace(t_root, &mut g)?;
+                        g_root = Some(g);
+                    }
+                    {
+                        let state = self.state_mut();
+                        state.y.copy_from(&y_root);
+                        *state.t = t_root;
+                        if let Some(g) = g_root.as_ref() {
+                            state.g.copy_from(g);
+                        }
+                    }
+                    write_out(self, &mut ret_y, &mut ret_t, &mut tmp_nout);
+                    break;
+                }
+            }
         }
-
-        // store the final step
-        write_out(self, &mut ret_y, &mut ret_t, &mut tmp_nout);
         let ntimes = ret_t.len();
         ret_y.resize_cols(ntimes);
         Ok((ret_y, ret_t))
@@ -202,6 +231,8 @@ where
     ///
     /// # Post-condition
     /// After the solver finishes, the internal state of the solver is at time `t_eval[t_eval.len()-1]`.
+    /// If a root is found, the solver stops early. The internal state is moved to the root time,
+    /// and the last column corresponds to the root time (which may not be in `t_eval`).
     fn solve_dense(
         &mut self,
         t_eval: &[Eqn::T],
@@ -214,14 +245,49 @@ where
 
         // do loop
         self.set_stop_time(t_eval[t_eval.len() - 1])?;
-        let mut step_reason = OdeSolverStopReason::InternalTimestep;
         for (i, t) in t_eval.iter().enumerate() {
             while self.state().t < *t {
-                step_reason = self.step()?;
+                match self.step()? {
+                    OdeSolverStopReason::InternalTimestep => {}
+                    OdeSolverStopReason::TstopReached => break,
+                    OdeSolverStopReason::RootFound(t_root) => {
+                        self.interpolate_inplace(t_root, &mut tmp_nstates)?;
+                        let integrate_out = self.problem().integrate_out;
+                        let mut g_root = None;
[…]
