### 2026-01-19 5b58b8d
doc: adjust line numbers in pred-prey fitting example
