### 2026-01-19 martinjrobins/diffsol#226 @martinjrobins
:( yea I keep running into this with the examples that use line numbers to include the examples. If you know of a better way to keep these in sync please let me know. For now I've just adjusted them manually in #234
