### 2026-01-19 30d5211
doc: update version num in readme
