### 2026-01-19 a88856e
doc: update version num in readme
