### 2026-01-19 df915d6
feat: add nonlinear solver tolerance and timestep growth/shrink user configuration options (#233)
