### 2026-01-19 ac92098
doc: correct dataset and add more docs for neural ode example
