### 2026-01-19 1dcafb1
doc: move module docs to mod files
