### 2026-01-19 0740668
Avoid Jacobian reset for initial and final checkpoints
