+++ diffsol/src/ode_solver/sdirk.rs
@@ -699,8 +699,8 @@ mod test {
         test_adjoint(adjoint_solver, dgdu);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
         number_of_calls: 419
-        number_of_jac_muls: 12
-        number_of_matrix_evals: 6
+        number_of_jac_muls: 10
+        number_of_matrix_evals: 5
         number_of_jac_adj_muls: 271
         "###);
     }
@@ -721,8 +721,8 @@ mod test {
         test_adjoint_sum_squares(adjoint_solver, dgdp, soln, data, times.as_slice());
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
         number_of_calls: 362
-        number_of_jac_muls: 12
-        number_of_matrix_evals: 4
+        number_of_jac_muls: 9
+        number_of_matrix_evals: 3
         number_of_jac_adj_muls: 686
         "###);
     }
@@ -740,8 +740,8 @@ mod test {
         test_adjoint(adjoint_solver, dgdu);
         insta::assert_yaml_snapshot!(problem.eqn.rhs().statistics(), @r###"
         number_of_calls: 433
-        number_of_jac_muls: 30
-        number_of_matrix_evals: 10
+        number_of_jac_muls: 27
+        number_of_matrix_evals: 9
         number_of_jac_adj_muls: 169
         "###);
     }
