### 2026-01-19 e575b26
perf: avoid Jacobian resets for start/endpoint checkpoints (#237)
* Avoid Jacobian reset for initial and final checkpoints

* update snapshots
