### 2026-01-20 376f9ee
fix run command and add matplotlib req
