### 2026-01-20 3ed7adc
dep: update diffsl 0.9.0 (#238)
* dep: update diffsl 0.9.0

* add new diffsl compiler options
