### 2026-01-21 41c8dcd
fix: dense_solve writes out all t_evals before root
