### 2026-01-22 #9384 @Armavica
Hi @nkrusch, I am nearing completion of the checklist, I just have a question about the **Performance** item: does it mean that I need to run the benchmarks (or some of them) shown in the documentation, comparing `diffsol` to preexisting software?
