### 2026-01-22 #9384 @nkrusch
@Armavica, the purpose is to check that the performance claims, that are reported in the paper, are consistent with the actual software behavior. However, if no performance claims are reported, the criteria is trivially satisfactory.

I have checked the paper for performance claims (typically: experiments with measurements). I did not spot any such claims in the paper. My opinion is that there are no performance claims to confirm.
