### 2026-01-22 9431f36
doc: correct dataset and improved docs for neural ode example (#235)
* remove generated data

* doc: correct dataset and add more docs for neural ode example

* fix run command and add matplotlib req

* fix clippy

* fix nn line numbers again
