+++ book/src/primer/weather_neural_ode.md
@@ -13,7 +13,7 @@ In this example, we will duplicate the weather prediction example from the excel
 First we'll need a neural network model, and we'll use [Equinox](https://github.com/patrick-kidger/equinox) and [JAX](https://docs.jax.dev/en/latest/index.html) for this. We'll define a simple neural network with 3 layers like so
 
 ```python
-{{#include ../../../examples/neural-ode-weather-prediction/src/model/model.py:39:54}}
+{{#include ../../../examples/neural-ode-weather-prediction/src/model/model.py:40:55}}
 ```
 
 We will then create four JAX functions that will allow us to calculate:
@@ -26,43 +26,43 @@ We will then create four JAX functions that will allow us to calculate:
 We will need all four of these to define the ODE problem and to solve it using Diffsol.
 
 ```python
-{{#include ../../../examples/neural-ode-weather-prediction/src/model/model.py:57:80}}
+{{#include ../../../examples/neural-ode-weather-prediction/src/model/model.py:66:81}}
 ```
 
 Finally, we can export all four of these JAX functions to ONNX, which will allow us to use them within rust.
 
 ```python
-{{#include ../../../examples/neural-ode-weather-prediction/src/model/model.py:25:33}}
+{{#include ../../../examples/neural-ode-weather-prediction/src/model/model.py:26:34}}
 ```
 
 Within rust now, we can define a Diffsol system of equations by creating a struct `NeuralOde`. We'll use the [`ort`](https://ort.pyke.io/) crate and the ONNX Runtime to load the ONNX models that we made in Python.
 
 ```rust,ignore
-{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:33:85}}
+{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:34:43}}
 ```
 
 We'll also implement the `OdeSystemAdjoint` trait for `NeuralOde`, which will allow us to use the adjoint method to calculate gradients of out loss function with respect to the parameters of the neural network. As an example, here is the implementation of the `NonLinearOp` trait:
 
 ```rust,ignore
-{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:196:220}}
+{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:206:228}}
 ```
 
 We'll also need an optimiser, so we'll write an AdamW algorithm using the definition in the [PyTorch documentation](https://pytorch.org/docs/stable/generated/torch.optim.AdamW.html) as a guide:
 
 ```rust,ignore
-{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:318:374}}
+{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:323:379}}
 ```
 
 We'll then define our loss function, which will return the sum of squared errors between the solution and the data points, along with the gradients of the loss function with respect to the parameters. Since the size of the parameter vector is quite large (>2000), we'll use the adjoint method to calculate the gradients.
 
 ```rust,ignore
-{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:376:397}}
+{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:381:402}}
 ```
 
 Finally, we can train the neural network to predict the weather. Following the example given in the linked blog post above, we'll train in stages by increasing the number of datapoints by four each time. Each time we'll train for 150 steps using the AdamW optimiser.
 
 ```rust,ignore
-{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:408:427}}
+{{#include ../../../examples/neural-ode-weather-prediction/src/main.rs:413:432}}
 ```
 
 To give an indication of the results, we'll plot the results after we've used the first 20 data-points to train the model, and we'll predict the model solution to the entire dataset.
