### 2026-01-23 83e9e0e
minor bib and paper edits
