+++ paper/paper.bib
@@ -1,7 +1,7 @@
 @software{cranelift,
   author = {{Bytecode Alliance}},
   title = {cranelift},
-  url = {https://wasmtime.dev/},
+  url = {https://wasmtime.dev},
   version = {0.115.1},
   date = {2025-03-13},
 }
@@ -17,7 +17,7 @@
 @software{faer,
     author = {Sarah El Kazdadi},
     title = {faer-rs},
-    url = {https://faer-rs.github.io/},
+    url = {https://faer-rs.github.io},
     version = {0.1.0},
     date = {2025-03-13},
 }
@@ -35,7 +35,7 @@
 }
 
 @article{shampine1997matlab,
-  title={The matlab ode suite},
+  title={The {MATLAB} {ODE} suite},
   author={Shampine, Lawrence F and Reichelt, Mark W},
   journal={SIAM journal on scientific computing},
   volume={18},
@@ -164,7 +164,8 @@ doi = {10.1109/SC41404.2022.00065}
 @software{PyO3_Project_and_Contributors,
 author = {{PyO3 Project and Contributors}},
 license = {["Apache-2.0", "MIT"]},
-url = {https://github.com/PyO3},
+url = {https://github.com/PyO3/pyo3},
+year = {2025},
 title = {{PyO3}}
 }
 
