### 2026-01-23 3c44086
add pyo3 version
