+++ paper/paper.bib
@@ -166,6 +166,7 @@ author = {{PyO3 Project and Contributors}},
 license = {["Apache-2.0", "MIT"]},
 url = {https://github.com/PyO3/pyo3},
 year = {2025},
+version = {v0.27.2},
 title = {{PyO3}}
 }
 
