### 2026-01-23 8f46c43
Minor JOSS paper + bib edits (#240)
* minor bib and paper edits

* add pyo3 version

Ref: JOSS #9384

---------

Co-authored-by: Martin Robinson <martinjrobins@gmail.com>
