### 2026-01-23 cae1d90
partially draft out burn example
