+++ examples/neural-ode-pk-dosing/src/main.rs
@@ -1,0 +1,7 @@
+mod rhs_network;
+mod equations;
+mod ode_backend;
+
+fn main() {
+    println!("Hello, world!");
+}
