+++ examples/neural-ode-pk-dosing/src/ode_backend.rs
@@ -1,0 +1,17 @@
+use burn::{backend::{NdArray, ndarray::{FloatNdArrayElement, NdArrayElement, NdArrayTensor}}, module::Module, prelude::Backend, tensor::ops::FloatTensor};
+use diffsol::Scalar;
+
+use crate::equations::Equations;
+
+
+trait OdeBackend: Backend {
+    fn solve<M: Module<Self>>(tensor: FloatTensor<Self>, rhs: &M) -> FloatTensor<Self>;
+}
+
+impl OdeBackend for NdArray {
+    fn solve<M: Module<Self>>(tensor: FloatTensor<Self>, rhs: &M) -> FloatTensor<Self> {
+        let eqn = Equations::new(rhs, &tensor);
+        // Implement the ODE solving logic here using the NdArray backend
+        tensor // Placeholder return
+    }
+}
