+++ examples/neural-ode-pk-dosing/Cargo.toml
@@ -1,0 +1,16 @@
+[package]
+name = "neural-ode-pk-dosing"
+version = "0.1.0"
+edition.workspace = true
+authors.workspace = true
+repository.workspace = true
+
+[dependencies]
+diffsol = { path = "../../diffsol" }
+nalgebra = { workspace = true }
+plotly = { workspace = true }
+csv = "1.3.1"
+burn = { version = "0.20.0", features = ["ndarray"] }
+
+[[bin]]
+name = "neural-ode-pk-dosing"
