### 2026-01-23 #9384 @martinjrobins
Thanks @nkrusch, @DiogoRibeiro7, @Armavica & @jeertmans for your help in reviewing the paper/software and providing very useful feedback :)

I've made a new release and deposited it on zendo, details are:

the version tag: v0.10.3
the archive DOI: https://zenodo.org/records/18351539
