### 2026-01-23 4fd9336
Fix DOI format in paper.bib (#241)
