+++ paper/paper.bib
@@ -77,7 +77,7 @@
     author={Patrick Kidger},
     year={2021},
     school={University of Oxford},
-    doi={h10.48550/arXiv.2202.02435}
+    doi={10.48550/arXiv.2202.02435}
 }
 
 @article{byrne1975polyalgorithm,
