### 2026-01-24 #9384 @danielskatz
As track editor, I'll work on this later today
