### 2026-01-24 c1d387f
case changes in paper.bib
