+++ paper/paper.bib
@@ -23,7 +23,7 @@
 }
 
 @article{virtanen2020scipy,
-  title={SciPy 1.0: fundamental algorithms for scientific computing in Python},
+  title={SciPy 1.0: fundamental algorithms for scientific computing in {P}ython},
   author={Virtanen, Pauli and Gommers, Ralf and Oliphant, Travis E and Haberland, Matt and Reddy, Tyler and Cournapeau, David and Burovski, Evgeni and Peterson, Pearu and Weckesser, Warren and Bright, Jonathan and others},
   journal={Nature methods},
   volume={17},
@@ -66,14 +66,14 @@
  note = {Exported from https://app.dimensions.ai on 2019/05/05},
  number = {1},
  pages = {},
- title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in Julia},
+ title = {DifferentialEquations.jl – A Performant and Feature-Rich Ecosystem for Solving Differential Equations in {J}ulia},
  url = {https://app.dimensions.ai/details/publication/pub.1085583166 and http://openresearchsoftware.metajnl.com/articles/10.5334/jors.151/galley/245/download/},
  volume = {5},
  year = {2017}
 }
 
 @phdthesis{kidger2021on,
-    title={{O}n {N}eural {D}ifferential {E}quations},
+    title={On Neural Differential Equations},
     author={Patrick Kidger},
     year={2021},
     school={University of Oxford},
@@ -150,7 +150,7 @@ doi = {10.1109/SC41404.2022.00065}
 }
 
 @article{tsitouras2011runge,
-  title={Runge--Kutta pairs of order 5 (4) satisfying only the first column simplifying assumption},
+  title={Runge--{K}utta pairs of order 5 (4) satisfying only the first column simplifying assumption},
   author={Tsitouras, Ch},
   journal={Computers \& mathematics with applications},
   volume={62},
@@ -176,7 +176,7 @@ doi = {10.4173/mic.2020.4.1},
 journal = {Modeling, Identification and Control},
 number = {4},
 pages = {241--295},
-title = {{The OpenModelica Integrated Environment for Modeling, Simulation, and Model-Based Development}},
+title = {The {OpenModelica} Integrated Environment for Modeling, Simulation, and Model-Based Development},
 volume = {41},
 year = {2020}
 }
