### 2026-01-24 #9384 @danielskatz
@martinjrobins - I've suggested the changes in https://github.com/martinjrobins/diffsol/pull/242 - please merge this, or let me know what you disagree with, then we can continue to publication
