### 2026-01-24 92dc853
minor changes for JOSS publication (#242)
* minor changes in paper.md

* case changes in paper.bib
