### 2026-01-24 #9384 @martinjrobins
thanks @danielskatz, these look good and I've merged them in
