### 2026-01-24 #9384 @editorialbot
:wave: @openjournals/csism-eics, this paper is ready to be accepted and published.
