### 2026-01-26 218542b
draft out neural ode example
