+++ examples/neural-ode-pk-dosing/src/equations.rs
@@ -1,43 +1,43 @@
-use burn::{Tensor, module::Module, prelude::*};
-use diffsol::{ConstantOp, ConstantOpSensAdjoint, NalgebraContext, NalgebraMat, NalgebraVec, NonLinearOp, NonLinearOpAdjoint, NonLinearOpSensAdjoint, OdeEquations, OdeEquationsRef, Op, Scalar, UnitCallable};
+use burn::{
+    tensor::{ops::FloatTensor, TensorPrimitive},
+    Tensor,
+};
+use diffsol::{
+    ConstantOp, ConstantOpSensAdjoint, DefaultDenseMatrix, NonLinearOp, NonLinearOpAdjoint,
+    NonLinearOpSensAdjoint, OdeEquations, OdeEquationsRef, Op, UnitCallable,
+};
+use num_traits::Zero;
 
-type T = f32;
-type V = NalgebraVec<T>;
-type M = NalgebraMat<T>;
-type C = NalgebraContext;
+use crate::{backend::Solve, rhs::Rhs, vector::Vector};
 
-pub struct Equations<'m, B, M> 
-where 
-    B: Backend,
-    M: Module<B>,
-{
+pub struct Equations<'m, V: Vector, B: Solve<V>, M: Rhs<B>> {
     rhs: &'m M,
-    y0: &'m Tensor<B, 1>,
-    input: &'m Tensor<B, 1>,
-    ctx: C,
+    y0: V,
+    params: FloatTensor<B>,
+    ctx: V::C,
+    _marker: std::marker::PhantomData<B>,
 }
 
-impl<'m, B, M> Equations<'m, B, M> 
-where 
-    B: Backend,
-    M: Module<B>,
-{
-    fn new(rhs: &'m M, y0: &'m Tensor<B, 1>, input: &'m Tensor<B, 1>) -> Self {
-        Self { rhs, y0, input, ctx: C::default() }
-    } 
+impl<'m, V: Vector, B: Solve<V>, M: Rhs<B>> Equations<'m, V, B, M> {
+    pub fn new(rhs: &'m M, y0: V, params: FloatTensor<B>) -> Self {
+        let ctx = y0.context().clone();
+        Self {
+            rhs,
+            y0,
+            params,
+            ctx,
+            _marker: std::marker::PhantomData,
+        }
+    }
     fn nstates(&self) -> usize {
-        self.y0.shape()[0]
+        self.y0.len()
     }
 }
-impl<'m, B, Mod> Op for Equations<'m, B, Mod> 
-where 
-    B: Backend,
-    Mod: Module<B>,
-{
-    type T = T;
+impl<'m, V: Vector, B: Solve<V>, Mod: Rhs<B>> Op for Equations<'m, V, B, Mod> {
+    type T = V::T;
     type V = V;
-    type M = M;
-    type C = C;
+    type M = <V as DefaultDenseMatrix>::M;
+    type C = V::C;
     fn nstates(&self) -> usize {
         self.nstates()
     }
@@ -52,25 +52,17 @@ where
     }
 }
 
-impl<'a, 'm, B, Mod> OdeEquationsRef<'a> for Equations<'m, B, Mod> 
-where 
-    B: Backend,
-    Mod: Module<B>,
-{
-    type Mass = UnitCallable<M>;
-    type Rhs = Rhs<'a>;
-    type Root = UnitCallable<M>;
-    type Init = Init<'a>;
-    type Out = UnitCallable<M>;
[…]
