+++ diffsol/src/lib.rs
@@ -346,12 +346,12 @@ use op::{
     closure_no_jac::ClosureNoJac, closure_with_sens::ClosureWithSens,
     constant_closure_with_sens::ConstantClosureWithSens, init::InitOp,
 };
-pub use scalar::{IndexType, Scalar, Scale};
+pub use scalar::{IndexType, Scalar, Scale, ScalarEnum};
 pub use vector::DefaultDenseMatrix;
 pub use vector::{
     faer_serial::{FaerVec, FaerVecIndex, FaerVecMut, FaerVecRef},
     nalgebra_serial::{NalgebraVec, NalgebraVecMut, NalgebraVecRef},
-    Vector, VectorCommon, VectorHost, VectorIndex, VectorRef, VectorView, VectorViewMut,
+    Vector, VectorCommon, VectorHost, VectorViewHost, VectorIndex, VectorRef, VectorView, VectorViewMut,
 };
 
 #[cfg(feature = "cuda")]
