+++ examples/neural-ode-pk-dosing/Cargo.toml
@@ -10,7 +10,7 @@ diffsol = { path = "../../diffsol" }
 nalgebra = { workspace = true }
 plotly = { workspace = true }
 csv = "1.3.1"
-burn = { version = "0.20.0", features = ["ndarray"] }
+burn = { version = "0.20.0", features = ["ndarray", "autodiff"] }
 num-traits = "0.2.19"
 ndarray = "0.17.2"
 
