+++ .github/workflows/gpu.yml
@@ -5,4 +5,4 @@ jobs:
     runs-on:
       - codebuild-diffsol-${{ github.run_id }}-${{ github.run_attempt }}
     steps:
-      - run: echo "Hello World 2"
+      - run: echo "Hello World 3"
