### 2026-01-29 #9384 @ChrisRackauckas
> both these packages compile the entire solver as well as the equations, which is a significant amount of code

This is inaccurate. The Julia package ahead of time compiles to a package binary which is then linked to via a FunctionWrapper, so it does not compile the whole solver and only compiles the equations when in the standard AutoSpecialize mode added in 2022. You can go back to the old mode via a flip to FullSpecialize, achieving higher performance via inter-procedural optimizations, but it has not been the default mode for quite a few years now.
