### 2026-01-29 #9384 @martinjrobins
Thanks @ChrisRackauckas, this is useful and I should have mentioned the AutoSpecialize option in the paper, my apologies. However, AutoSpecialize seems to be not applicable in many of the usecases that diffsol targets. I've been reading the docs [here](https://docs.sciml.ai/SciMLBase/stable/interfaces/Problems/#Specialization-Levels), these seem to indicate that  it is only meant for REPL use and not recommended for use when optimal runtime performance of the solver is required (e.g. optimisation loops), and cannot be used at all when (a) a jacobian is supplied, (b) with non-singular mass matrices (i.e. DAEs), or (c) when calculating sensitivities? For these cases, can you clarify if the solver will be JIT compiled or ahead-of-time compiled?
