### 2026-01-29 6fead51
refactor: remove DefaultSolver bound from solve_dense_sensitivities
