### 2026-01-30 #9384 @ChrisRackauckas
That seems to be mixing up "auto" with possible. We automatically choose between sometimes compiling all and sometimes using function wrappers to reduce latency. That depends on whether a solve is going to take a long time or not. If it's going to take minutes, like something hitting optimization with linear solvers, why would you try to save 20 seconds on compile to make it run 3 minutes longer? Of course the heuristics can miss so you can FullSpecialize or FunctionWrapperSpecialize by choice, but auto is generally what the user wants.

So yes, all cases you specify can be AOT compiled, it's just a question whether you want to, and we don't always flip the switch automatically with the defaults because it's not always a good idea.
